% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac_convert_units}
\alias{ac_convert_units}
\title{Convert values from one unit to another using database-defined conversions
This function retrieves the conversion parameters from the database and applies the appropriate transformation to convert values from the specified 'from_unit' to the 'to_unit'. If no conversion is found and 'strict' is TRUE, an error is raised. If 'strict' is FALSE, the original values are returned unchanged.}
\usage{
ac_convert_units(con, values, from_unit, to_unit, strict = TRUE)
}
\arguments{
\item{con}{A DBI connection to the database containing unit conversion information.}

\item{values}{A numeric vector of values to be converted.}

\item{from_unit}{A string representing the unit of the input values.}

\item{to_unit}{A string representing the desired unit for the output values.}

\item{strict}{A logical value indicating whether to raise an error if no conversion is found (default is TRUE). If FALSE, the original values will be returned unchanged when no conversion is found.}
}
\value{
A numeric vector of converted values, or the original values if no conversion is found and 'strict' is FALSE.
}
\description{
Convert values from one unit to another using database-defined conversions
This function retrieves the conversion parameters from the database and applies the appropriate transformation to convert values from the specified 'from_unit' to the 'to_unit'. If no conversion is found and 'strict' is TRUE, an error is raised. If 'strict' is FALSE, the original values are returned unchanged.
}
