% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac_get_parameter_unit}
\alias{ac_get_parameter_unit}
\title{Get the unit for a parameter from the database
Retrieves the unit associated with a given parameter_id from the database, using the specified preference for which unit to return if multiple are available. The preference can be set to prioritize the default unit, the solid unit, or a combination of both. If no unit is found, returns NA.}
\usage{
ac_get_parameter_unit(
  con,
  parameter_id,
  prefer = c("default_or_solid", "default", "solid", "solid_or_default")
)
}
\arguments{
\item{con}{A DBI connection to the database containing parameter and unit information.}

\item{parameter_id}{An integer representing the parameter_id for which to retrieve the unit.}

\item{prefer}{A string indicating the preference for which unit to return if multiple are available. Options are "default_or_solid" (default), "default", "solid", and "solid_or_default".}
}
\value{
A string representing the unit associated with the specified parameter_id, or NA if no unit is found.
}
\description{
Get the unit for a parameter from the database
Retrieves the unit associated with a given parameter_id from the database, using the specified preference for which unit to return if multiple are available. The preference can be set to prioritize the default unit, the solid unit, or a combination of both. If no unit is found, returns NA.
}
