% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac_parameter_unit_select_sql}
\alias{ac_parameter_unit_select_sql}
\title{Generate SQL expression to select parameter unit with fallback
Generates a SQL expression to select the unit for a parameter, with fallback logic based on the available columns in the database. The function checks for the presence of unit columns in the 'parameters' table and constructs a SQL expression that retrieves the unit name, either directly from text columns or by joining with the 'units' table if normalized unit ID columns are present. The preference for which unit to return can be specified through the 'prefer' argument.}
\usage{
ac_parameter_unit_select_sql(
  con,
  parameter_alias = NULL,
  output_alias = "unit_default",
  prefer = c("default_or_solid", "default", "solid", "solid_or_default")
)
}
\arguments{
\item{con}{A DBI connection to the database containing parameter and unit information.}

\item{parameter_alias}{An optional string representing the table alias for the parameters table in the SQL query. If provided, column references will be prefixed with this alias.}

\item{output_alias}{A string representing the desired alias for the output column in the SQL query. Defaults to "unit_default".}

\item{prefer}{A string indicating the preference for which unit to return if multiple are available. Options are "default_or_solid" (default), "default", "solid", and "solid_or_default".}
}
\value{
A string containing the SQL expression to select the parameter unit with appropriate fallback logic based on the database schema.
}
\description{
Generate SQL expression to select parameter unit with fallback
Generates a SQL expression to select the unit for a parameter, with fallback logic based on the available columns in the database. The function checks for the presence of unit columns in the 'parameters' table and constructs a SQL expression that retrieves the unit name, either directly from text columns or by joining with the 'units' table if normalized unit ID columns are present. The preference for which unit to return can be specified through the 'prefer' argument.
}
