% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_download.R
\name{aq_download}
\alias{aq_download}
\title{Get formatted timeseries data from Aquarius.}
\usage{
aq_download(
  loc_id,
  ts_name,
  start = "1950-01-01",
  end = Sys.Date(),
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = Sys.getenv("AQSERVER")
)
}
\arguments{
\item{loc_id}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}.}

\item{ts_name}{The timeseries name, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \code{Wlevel_bgs.Calculated}.}

\item{start}{The first day or instant for which you want information, in UTC 0 timezone. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the first moment of the day. Times requested prior to the actual timeseries start will be adjusted to match available data.}

\item{end}{The last day or instant for which you want information, in UTC 0. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the last moment of the day. Times requested prior to the actual timeseries end will be adjusted to match available data.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron file; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default pulls from your .renviron file; see details.}
}
\value{
A list with four data.frames: station metadata; timeseries information consisting of datetimes, values, applicable grade and approval levels; approval level change summary; grade level change summary. IMPORTANT: all times in this returned list are in UTC.
}
\description{
Fetches and processes data downloaded from an Aquarius web-hosted server and returns it in a concise format. Note that any times returned by this function are in UTC.
}
\details{
To store login credentials in your .renviron file, call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS="your password". The server should be entered at server="your_server_url".
}
