% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_download_curves.R
\name{aq_download_curves}
\alias{aq_download_curves}
\title{Get flow rating curves from Aquarius}
\usage{
aq_download_curves(
  loc_id,
  start = "1950-01-01",
  end = Sys.Date(),
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = Sys.getenv("AQSERVER")
)
}
\arguments{
\item{loc_id}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}. Note that the timeseries id is not used, only the location ID.}

\item{start}{The first day or instant for which you want information, in UTC 0 timezone. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the first moment of the day. Times requested prior to the actual timeseries start will be adjusted to match available data.}

\item{end}{The last day or instant for which you want information, in UTC 0. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the last moment of the day. Times requested prior to the actual timeseries end will be adjusted to match available data.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron file; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default pulls from your .renviron file; see details.}
}
\value{
A list with two data.frames: station metadata (simple data.frame) and rating curve information (complex data.frame). The later contains one row per parameter which has a rating curve (i.e. a distinct curve ID), with the final column containing a list of all the curves associated with that curve ID.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Fetches and processes rating curve data downloaded from an Aquarius web-hosted server and returns it in a concise format. Note that any times returned by this function are in UTC.
}
\details{
To store login credentials in your .renviron file, call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS="your password". The server should be entered at server="your_server_url".
}
