% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{AquaConnect}
\alias{AquaConnect}
\title{Connect to the AquaCache database}
\usage{
AquaConnect(
  name = "AquaCache",
  host = Sys.getenv("AquaCacheHost"),
  port = Sys.getenv("AquaCachePort"),
  username = Sys.getenv("AquaCacheUser"),
  password = Sys.getenv("AquaCachePass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form AquaCacheHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter=value pair of form AquaCachePort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter=value pair of form AquaCacheUser="username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter=value pair of form AquaCachePass="password".}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the production version AquaCache database. For the development version see \code{\link[=hydrometConnect]{hydrometConnect()}}.
}
