% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaLink.R
\name{AquaLink}
\alias{AquaLink}
\title{Launch the AquaLink Shiny application}
\usage{
AquaLink(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port"),
  dbName = Sys.getenv("aquacacheName", "aquacache"),
  dbHost = Sys.getenv("aquacacheHost", "localhost"),
  dbPort = Sys.getenv("aquacachePort", "5432"),
  dbUser = Sys.getenv("aquacacheUser"),
  dbPass = Sys.getenv("aquacachePass"),
  accessPaths =
    c("//env-fs/env-data/corp/water/Data/Databases_virtual_machines/databases/EQWinDB",
    "//carver/infosys/EQWin"),
  network_check = FALSE,
  logout_timer_min = 10,
  server = FALSE,
  analytics = FALSE,
  public = TRUE,
  brand = "yukon"
)
}
\arguments{
\item{host}{Host address. Leave default to run locally, set to "0.0.0.0" to enable others to connect. Depends on the port specified in \code{port} to be open on the host machine. Ignored if \code{server} is set to TRUE (Shiny Server takes care of that).}

\item{port}{Port number (numeric) on which to serve the app. Leave default to use the default port specified in your user options. The port you specify must be open on the host machine for it to broadcast to the network. Ignored if \code{server} is set to TRUE (Shiny Server takes care of that).}

\item{dbName}{Name of the aquacache database. Searches your environment variables for 'aquacacheName', default is "aquacache" if not found.}

\item{dbHost}{Host address of the aquacache database. Default is pulled from the .Renviron file.}

\item{dbPort}{Port number of the aquacache database. Default is pulled from the .Renviron file.}

\item{dbUser}{Username for the aquacache database. Default is pulled from the .Renviron file.}

\item{dbPass}{Password for the aquacache database. Default is pulled from the .Renviron file.}

\item{accessPaths}{Paths to the folders where EQWin databases are stored. Default is "//env-fs/env-data/corp/water/Data/Databases_virtual_machines/databases/EQWinDB". The function will search for all *.mdb files in this folder (but not its sub-folders) and list them as options.}

\item{network_check}{Optional: a path to a network location to check if the app is run internally or externally. If the app can see the path, some additional functionality is enabled, such as log in by admin roles. Default is FALSE, which means no check is performed and the app assumes it's running internally without access to all features.}

\item{logout_timer_min}{Auto logout timer, in minutes.}

\item{server}{Set to TRUE to run on Shiny Server, otherwise FALSE to run locally.}

\item{analytics}{FALSE (no analytics) or path to an HTML file containing analytics code to insert into the app, such as a matomo analytics script. If a file in the package's inst directory, point to it as \code{system.file("apps/aqualink/www/html/matomo.html", package = "YGwater")}. This file will be called within the main UI as \code{tags$head(includeHTML(file))}. Default is FALSE.}

\item{public}{TRUE restricts or doesn't create some UI elements that are not intended for public use, such as a login button and some crude report generation modules. Default is TRUE.}

\item{brand}{Character string to specify the branding of the app. Default is "public", which is a more generic version of the app with Yukon-specific references removed. The "yukon" branding includes Yukon-specific logos and references. This parameter can be used in the future to create other branded versions of the app for different users or organizations.}
}
\value{
Opens a Shiny application.
}
\description{
Launches a Shiny application to visualize the WRB's water-related data in maps and plots, and to download data from the WRB's hydrometric database. The application is designed to work with the WRB's aquacache database, which contains hydrometric data, and optionally with EQWin databases for water quality data.
}
