% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeReport.R
\name{bridgeReport}
\alias{bridgeReport}
\title{Bridge radar reporting utility - internal}
\usage{
bridgeReport(
  con = AquaConnect(silent = TRUE),
  locations = "all",
  zoom = TRUE,
  zoom_days = 30,
  save_path = "choose"
)
}
\arguments{
\item{con}{A connection to the database. Default uses function \code{\link[=AquaConnect]{AquaConnect()}} with default settings.}

\item{locations}{The list of locations for which you want a distance measurement, or default "all" to get every one in the database. These must be reporting radar distance in the WRB database as parameter 'distance'. Default "all" will only fetch locations where the network is listed as 'Highway Observation Network' in the locations_networks table of the database.}

\item{zoom}{Set TRUE if you want zoomed-in plots.}

\item{zoom_days}{Set the number of days on the x-axis of the zoomed in plots.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default "choose" lets you select your folder, otherwise enter the path as a character string. WARNING: option 'choose' only works on Windows, and some late-build R versions have a bug that prevents it from working every time.}
}
\value{
A Microsoft Word report containing bridge radar information, saved in the location specified.
}
\description{
This function generates a report of distance between the water surface and bridges or other important infrastructure. The output is a Microsoft Word document on a Yukon Government template.
}
