% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{calculate_fdd}
\alias{calculate_fdd}
\title{Calculate historical norms for stations}
\usage{
calculate_fdd(
  temperature,
  percent_missing_threshold = 0.8,
  water_year_start_month = 10
)
}
\arguments{
\item{percent_missing_threshold}{Numeric (0-1) minimum data completeness in cumulative FDD calculation}

\item{water_year_start_month}{Integer month when water year starts (default October = 10)}

\item{temperature_data}{data.frame with 'datetime' column and station columns}
}
\value{
data.frame with cumulative FDD for each station
}
\description{
Calculate historical norms for stations
}
