% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-helpers.R
\name{check_conda_package_installed}
\alias{check_conda_package_installed}
\title{Check if a specific package is installed in a conda environment}
\usage{
check_conda_package_installed(env_name, package_name)
}
\arguments{
\item{env_name}{The name of the version of Python to be used by \code{reticulate}.}

\item{package_name}{The name of the caonda package to check for.}
}
\value{
TRUE if installed, FALSE if not.
}
\description{
Check if a specific package is installed in a conda environment
}
