% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-helpers.R
\name{check_miniconda_installed}
\alias{check_miniconda_installed}
\title{Check if miniconda is installed}
\usage{
check_miniconda_installed()
}
\value{
Nothing.
}
\description{
Checks for a miniconda installation and installs it if necessary using reticulate::install_miniconda.
}
