% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-helpers.R
\name{check_reticulate}
\alias{check_reticulate}
\title{Check if reticulate package is installed}
\usage{
check_reticulate(silent)
}
\arguments{
\item{silent}{If TRUE, reticulate is installed if missing without any user input.}
}
\value{
Nothing.
}
\description{
Check if reticulate package is installed
}
