% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cached}
\alias{clear_cached}
\title{Remove a value from the cache}
\usage{
clear_cached(key, env = .GlobalEnv)
}
\arguments{
\item{key}{The cached key to remove.}

\item{env}{The environment from which to remove the cached key. Defaults to \code{.GlobalEnv}.}
}
\description{
Remove a value from the cache
}
