% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-helpers.R
\name{conda_installed_packages}
\alias{conda_installed_packages}
\title{Retrieve installed packages in a conda environment}
\usage{
conda_installed_packages(env_name)
}
\arguments{
\item{env_name}{The name of the version of Python to be used by \code{reticulate}.}
}
\value{
The installed packages in the specified conda environment.
}
\description{
Retrieve installed packages in a conda environment
}
