% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{create_survey_summary}
\alias{create_survey_summary}
\title{create a summary of snow survey stations with their associated basins, to be used for the snow bulletin appendix}
\usage{
create_survey_summary(con, year = 2025, month = 3, epsg = 4326)
}
\arguments{
\item{con}{DBI database connection object}

\item{year}{Integer year for which to load survey data (default 2025)}

\item{month}{Integer month for which to load survey data (default 3)}

\item{epsg}{Integer EPSG code for coordinate reference system (default 4326)}
}
\value{
a table containing summary of snow survey stations with their associated basins
}
\description{
This function queries the database for snow survey station locations, performs a spatial join with basin shapefiles to assign each station to a basin, and compiles a summary table of stations and their associated basins. This summary can be used for documentation purposes in the snow bulletin appendix, providing an overview of the survey network and its coverage across different basins.
}
