% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSnowOpenData.R
\name{createSnowOpenData}
\alias{createSnowOpenData}
\title{Creates the files for download on Yukon Open Data}
\usage{
createSnowOpenData()
}
\value{
Two excel documents, one with the locations, another with the data. The snow data includes all surveys in the database. it does not include individual samples, but simply a mean of all samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
The purpose of this script is create files for the Yukon Open Data portal.
}
