% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSnowTemplate.R
\name{createSnowTemplate}
\alias{createSnowTemplate}
\title{Creates snow survey template for specific survey target date and circuit}
\usage{
createSnowTemplate(target_date, circuit = "all", save_path = "choose")
}
\arguments{
\item{target_date}{The target date of the snow survey, given in the form yyyy-mm-dd as a text string or Date object. Example: for the march snow survey of 2024, target_date = '2024-03-01'}

\item{circuit}{The circuit for which we are creating a snow survey template. Options are Carmacks, Dawson, HJ, KluaneP, Mayo, NorthSlope, OldCrow, PellyFarm, Ross, SLakes, Teslin, Watson, Whitehorse, and YEC. "all" will create a workbook for all circuits.}

\item{save_path}{The path where you want the circuit workbook(s) saved. Default "choose" lets you pick your folder.}
}
\value{
A snow survey template called template_test (currently) for the specified circuit and target date. This excel workbook has a sheet for every snow course within the circuit and contains a summary sheet with current and previous years stats.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function creates snow survey templates for a snow survey circuit in an excel workbook. The function takes the master template and copies it into multiple sheets with circuit name, course name and snow survey target date pre-filled. The outputted template is meant to be filled out by those conducting snow surveys.

You should be running this function from a computer with access to the Snow database for full functionality:
}
