% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_browse_ts.R
\name{DB_browse_ts}
\alias{DB_browse_ts}
\title{Get information on timeseries contained in the database}
\usage{
DB_browse_ts(
  path = "default",
  operator = NULL,
  location = NULL,
  type = NULL,
  parameter = NULL
)
}
\arguments{
\item{path}{The path to the database, passed to \code{\link[=hydrometConnect]{hydrometConnect()}}. Default uses hydrometConnect default path.}

\item{operator}{Narrow by location operator if you wish (one or more). Exact spelling only!}

\item{location}{Narrow by location if you wish (one or more). Exact spelling only!}

\item{type}{Narrow by type if you wish (one or more), such as "discrete" or "continuous". Exact spelling only!}

\item{parameter}{Narrow by parameter if you wish (one or more), such as "level", "flow", "SWE"... Exact spelling only!}
}
\value{
A list of two data.frames: one containing the database timeseries and one containing location-specific information such as full names, coordinates. Both data.frames are filtered to match the parameters passed to the function
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Wondering what's in the database? This function helps you see what's under the hood, with an eye to helping you create a query for function \code{\link[=DB_get_ts]{DB_get_ts()}}. Leaving all NULL defaults will show you every timeseries in the database and additional data specific to each location.
}
\seealso{
\code{\link[=DB_get_ts]{DB_get_ts()}} for extracting timeseries data from the database, or, for spatial data, \code{\link[=DB_browse_spatial]{DB_browse_spatial()}} and \code{\link[=DB_get_spatial]{DB_get_spatial()}} to browse and extract spatial data. For location metadata use \code{\link[=DB_get_meta]{DB_get_meta()}}.
}
