% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_get_meta.R
\name{DB_get_meta}
\alias{DB_get_meta}
\title{Get location metadata from the hydromet database}
\usage{
DB_get_meta(con = hydrometConnect(), location = "all", save_path = NULL)
}
\arguments{
\item{con}{A connection to the database. Default uses function \code{\link[=hydrometConnect]{hydrometConnect()}} with default settings.}

\item{location}{A character vector of 1 or more location IDs, or default "all" to retrieval all locations.}

\item{save_path}{Specify a path here if you want an Excel workbook saved to disk. "choose" lets you interactively choose your folder.}
}
\value{
A list of two data.frames: one with the location code, name, latitude, and longitude, and another with vertical datum information. Optionally, an Excel workbook saved to disk.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieves metadata for any valid location ID in the database. Returns the name of the location, the lat/long and (if possible) vertical datum information.
}
\seealso{
\code{\link[=DB_browse_ts]{DB_browse_ts()}} and \code{\link[=DB_get_ts]{DB_get_ts()}} to browse for and extract timeseries data; \code{\link[=DB_browse_spatial]{DB_browse_spatial()}} and \code{\link[=DB_get_spatial]{DB_get_spatial()}} to browse and extract spatial layers.
}
