% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_get_spatial.R
\name{DB_get_spatial}
\alias{DB_get_spatial}
\title{Get spatial files from the database}
\usage{
DB_get_spatial(
  path = "default",
  type,
  rowid,
  save_path = NULL,
  save_name = NULL
)
}
\arguments{
\item{path}{The path to the database, passed to \code{\link[=hydroConnect]{hydroConnect()}}. Default uses hydroConnect default path.}

\item{type}{'polygon' or 'raster'?}

\item{rowid}{The rowid of the file you wish to download. If unsure, use function \code{\link[=DB_browse_spatial]{DB_browse_spatial()}} to narrow your search down to a single file.}

\item{save_path}{Optional; the path where the raster (as tif) or polygon (as shapefile) should be saved. You can enter 'choose' to select the path interactively.}

\item{save_name}{Optional, the name of the layer written to disk (without extension). Leave NULL to have it named vector or raster with the system date.}
}
\value{
A spatial object as per the terra package and, if requested, a shapefile or tif saved to disc.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Loads a database spatial file to the R environment and, optionally, saves the file to disc. Refer to function \code{\link[=DB_browse_spatial]{DB_browse_spatial()}} to identify the exact record you are looking for.
}
\details{
Spatial files are not stored directly in the database but rather in folders situated alongside the database. The database stores the file description and other identifiers, as well as the path to the file.
}
\seealso{
\code{\link[=DB_browse_spatial]{DB_browse_spatial()}} if you need to identify the exact record you're after.
}
