% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_crosstab.R
\name{EQ_crosstab}
\alias{EQ_crosstab}
\title{Concatenate 'crosstab' type lab reports for entry to EQWin}
\usage{
EQ_crosstab(folder = "choose", save_path = NULL)
}
\arguments{
\item{folder}{The folder containing the lab reports. Default "choose" opens a dialog box to select the folder.}

\item{save_path}{The path to save the concatenated report (folder path only). Default NULL saves the output in the folder provided in the first argument, 'choose' lets you choose interactively, or specify your own.}
}
\value{
An Excel file with the concatenated lab reports.
}
\description{
This function is designed to concatenate lab reports that are in a 'crosstab' format. The function reads in the lab reports, concatenates them, and saves the output as an Excel file. The function is designed to work with lab reports that are in a specific format, and may need to be modified to work with other formats.
Total phosphate measurements are sometimes reported twice under the same paramter name and code, with a result for the total phosphate and a result for the orthophosphate. This function will rename the second-listed total phosphate measurement to 'P-PO4T' to avoid duplicate column names. If this is NOT what you need, modify the parameter code yourself in the workbook before using this function.
}
