% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_fetch.R
\name{eq_fetch}
\alias{eq_fetch}
\title{Data retrieval from EQWin}
\usage{
eq_fetch(
  EQcode,
  stationIDs = "all",
  paramIDs = "all",
  dates = "all",
  BD = 2,
  apply_standards = TRUE,
  path = eqwin_db_path()
)
}
\arguments{
\item{EQcode}{Site code as it appears in EQWin eg. "LOB" or "KNO". Function only works for stations with  designated project code in brackets}

\item{stationIDs}{"all" for all stations (default) OR character vector of selected stations as they appear in the EQWin database WITHOUT the EQcode c("MW-01", "MW-02)}

\item{paramIDs}{"all" for all parameters (default) OR vector of selected parameters exactly as they appear in the EQWin database}

\item{dates}{"all" for all dates (default) OR character vector of length 2 of start and end date in format c("YYYY-MM-DD", "YYYY-MM-DD")}

\item{BD}{Treatment of values below detection limits (0 = Set to zero; 1 = Set to NA; 2 = Set to 0.5*(LOD); 3 = Set to sqrt(2)LOD). Above detection values are set to the upper limit of detection.}

\item{apply_standards}{TRUE or FALSE, include standards with data. Provides a pop-up list for selection.}

\item{path}{The path to the EQWin database.}
}
\value{
A list with one sub-list per station, each one containing 2 data frames with sample data and calculated standards
}
\description{
Fetches sample data from the WRB database and returns a list of data frames suitable for modification for plot generation and other comparisons. Connection to the EQWin database is made via function \code{\link[=AccessConnect]{AccessConnect()}}.
}
\details{
Insert here what happens to values > DL, where the standards are taken from, etc
}
