% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQchecks.R
\name{EQchecks}
\alias{EQchecks}
\title{Shiny application to facilitate QA/QC checks on EQWin data}
\usage{
EQchecks(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port")
)
}
\arguments{
\item{host}{Host address. Leave default to run locally, set to "0.0.0.0" to enable others to connect. Depends on the port specified in \code{port} to be open on the host machine.}

\item{port}{Port number (numeric). Leave default to use the default port specified in your user options. The port you specify must be open on the host machine for it to broadcast to the network.}
}
\value{
Opens a Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Allows the user to perform automatic check on EQWin samples within a user-defined date range and, optionally, at certain locations or for certain sampleIds only.
}
