% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQWinData.R
\name{EQWinData}
\alias{EQWinData}
\title{Extract EQWin data to an Excel workbook or .csv}
\usage{
EQWinData(
  start,
  end = Sys.Date() + 1,
  stations = NULL,
  stnGrp = NULL,
  parameters = NULL,
  paramGrp = NULL,
  format = "wide",
  DL1 = NULL,
  DL2 = NULL,
  stds = NULL,
  stnStds = TRUE,
  save_path = "choose",
  dbPath = eqwin_db_path()
)
}
\arguments{
\item{start}{The date to fetch data from, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'.}

\item{end}{The end date to fetch data up to, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'. Default is the current date.}

\item{stations}{A vector of station names as listed in the EQWiN eqstns table, column StnCode. Leave NULL to use stnGrp instead.}

\item{stnGrp}{A station group as listed in the EWQin eqgroups table, column groupname. Leave NULL to use stations instead.}

\item{parameters}{A vector of parameter names as listed in the EQWin eqparams table, column ParamCode. Leave NULL to use paramGrp instead.}

\item{paramGrp}{A parameter group as listed in the EQWin eqgroups table, column groupname. Leave NULL to use parameters instead.}

\item{format}{'long' or 'wide' to output the data in a long or wide format. Default is 'wide', in which locations are separated in workbook tabs, sample datetimes in columns, and parameters in rows. Long format is a single table with columns for 'location', 'parameter', and 'value'. Long format is useful for programmatic analysis and plotting, while wide format is useful for looking at the data and making plots in Excel.}

\item{DL1}{Behavior with respect to values listed as < the detection limit. Default is NULL which leaves the values as is; 'negative' replaces '<' with '-' and converts to numeric; 'half' replaces with 1/2 the detection limit and converts to numeric; 'NA' replaces the value with 'NA' and converts the rest to numeric.}

\item{DL2}{Behavior with respect to values listed as > the detection limit. Default is NULL which leaves the values as is; 'convert' simple removes the '<' and converts to numeric; 'NA' replaces the value with 'NA' and converts the rest to numeric.}

\item{stds}{A vector of standard names as listed in the EQWin eqstds table. Leave NULL to exclude standards. Standards will be output to an Excel workbook, one labelled tab per standard.}

\item{stnStds}{TRUE/FALSE to include/exclude the station-specific standards listed in the eqstns table, column StnStd. Station standards will be added to the 'standards' workbook as per parameter \code{stds}.}

\item{save_path}{The path to save the Excel file(s). Default is "choose" to allow user to select a folder interactively.}

\item{dbPath}{The path to the EQWin database. Defaults to a WRB network location.}
}
\value{
A list of data.frames containing the data, location and parameter info, and (optionally) standards, plus an Excel workbook containing the requested station data, an optional workbook containing the standards.
}
\description{
This function extracts data from the EQWin database and outputs it to an Excel workbook. The function can extract data from a specified date range, station(s), and parameter(s). The function can also output standard values as well as station-specific standards. .
}
\details{
Connection to EQWin is made using function \code{\link[=AccessConnect]{AccessConnect()}}.
}
\examples{
\dontrun{
# Using station and parameter groups
EQWinData(start = "2024-07-01 00:00", end = Sys.Date(), stnGrp = "QZ Eagle Gold HLF",
paramGrp = "EG-HLF-failure", format = 'wide', stds = c("CCME_LT", "CCME_ST"), stnStds = TRUE)

With specific stations and parameters
EQWinData(start = "2024-01-01 00:00", end = Sys.Date(), stations = c("(EG)W23"),
parameters = c("pH-F"), format = 'wide', stds = c("CCME_LT", "CCME_ST"), stnStds = TRUE)
}
}
