% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQWinPlot.R
\name{EQWinPlot}
\alias{EQWinPlot}
\title{EQWin plotting function}
\usage{
EQWinPlot(
  start,
  end = Sys.Date() + 1,
  stations = NULL,
  stnGrp = NULL,
  parameters = NULL,
  paramGrp = NULL,
  log = FALSE,
  title = TRUE,
  facet_on = "stns",
  save_path = "choose",
  dbPath = "//carver/infosys/EQWin/WR/DB/Water Resources.mdb"
)
}
\arguments{
\item{start}{The date to fetch data from, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'.}

\item{end}{The end date to fetch data up to, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'. Default is the current date.}

\item{stations}{A vector of station names as listed in the EQWiN eqstns table, column StnCode. Leave NULL to use stnGrp instead.}

\item{stnGrp}{A station group as listed in the EWQin eqgroups table, column groupname. Leave NULL to use stations instead.}

\item{parameters}{A vector of parameter names as listed in the EQWin eqparams table, column ParamCode. Leave NULL to use paramGrp instead.}

\item{paramGrp}{A parameter group as listed in the EQWin eqgroups table, column groupname. Leave NULL to use parameters instead.}

\item{log}{Should the y-axis be log-transformed?}

\item{title}{Should the plot have a title?}

\item{facet_on}{Should the plot be faceted by stations or by parameters?}

\item{save_path}{The path to save the plot as an html file. Default is "choose" to allow user to select a folder interactively.}

\item{dbPath}{The path to the EQWin database. Default is "//carver/infosys/EQWin/WR/DB/Water Resources.mdb".}
}
\value{
A zoomable plot of the data from EQWin.
}
\description{
Plots data directly from EQWin for one or more location (station) and one or more parameter. Depending on the setting for argument 'facet_on', the function can either make a facet plot where each station is a facet (with parameters as distinct traces) or where each parameter is a facet (with stations as distinct traces). Values above or below the detection limit are shown as the detection limit but symbolized differently. Duplicate samples are not shown.
}
