% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQWinReport.R
\name{EQWinReport}
\alias{EQWinReport}
\title{Create an Excel workbook report of water quality data from the EQWin database}
\usage{
EQWinReport(
  date,
  stations = NULL,
  stnGrp = NULL,
  parameters = NULL,
  paramGrp = NULL,
  stds = NULL,
  stnStds = TRUE,
  date_approx = 0,
  save_path = "choose",
  dbPath = "X:/EQWin/WR/DB/Water Resources.mdb"
)
}
\arguments{
\item{date}{The date for which to fetch results as a Date object or character vector that can be coerced to a date.}

\item{stations}{A vector of station names as listed in the EQWiN eqstns table, column StnCode. Leave NULL to use stnGrp instead.}

\item{stnGrp}{A station group as listed in the EWQin eqgroups table, column groupname. Leave NULL to use stations instead.}

\item{parameters}{A vector of parameter names as listed in the EQWin eqparams table, column ParamCode. Leave NULL to use paramGrp instead.}

\item{paramGrp}{A parameter group as listed in the EQWin eqgroups table, column groupname. Leave NULL to use parameters instead.}

\item{stds}{A vector of standard names as listed in the EQWin eqstds table. Leave NULL to exclude standards. As these can apply to all stations standard values will be listed in a column to the left of the results.}

\item{stnStds}{TRUE/FALSE to include/exclude the station-specific standards listed in the eqstns table, column StnStd. As these are station-specific, the standard values will be listed in a comment linked to the station name in the table header.}

\item{date_approx}{An optional maximum number of days in the past or future to fetch results for stations where data is unavailable for the exact date. If a station is found to not have any samples on the given 'date', the function will look for samples on the date + 1 day, - 1 day, + 2 days, etc. up to 'date_approx' days in the past and future and stop searching for a station once a sample is found. Default is 0.}

\item{save_path}{The path to save the Excel file. Default is "choose" to allow user to select a folder interactively.}

\item{dbPath}{The path to the EQWin database. Default is "X:/EQWin/WR/DB/Water Resources.mdb".}
}
\value{
An Excel workbook saved where requested.
}
\description{
Generates a report of water quality data for a given date, set of stations (or EQWin specified station group), and parameters (or specified parameter group). Guidelines passed to argument 'stds' are included as columns in the table, while station-specific guidelines are listed in a note in the cell containing the station name.

Exceedances are noted using conditional formatting, with a note added to explain which standard(s) are exceeded.
}
\details{
Connection to EQWin is made using function \code{\link[=AccessConnect]{AccessConnect()}}.
}
\examples{
\dontrun{

# These examples require access to the G drive to work and must write to a file. They are 
# provided for demonstration purposes only and can't be run using the 'Run examples' button above.
# Copy/paste them into your R console to run them if needed.

# Generate a report for July 1, 2024 using a station group, parameter group, 
# both CCME standards, and look for data within +- 1 day of July 1
EQWinReport("2024-07-25", stnGrp = "QZ Eagle Gold HLF", paramGrp = "EG-HLF-failure",
stds = c("CCME_ST", "CCME_LT"), date_approx = 1)


# Generate a report for a single location and single parameter, no standards.
EQWinReport("1991-02-12", stations = c("(CM)CM-u/s"), parameters = c("pH-F"), stnStds = FALSE)

}
}
