% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQWinStd.R
\name{EQWinStd}
\alias{EQWinStd}
\title{Calculate standards using formulas in EQWin}
\usage{
EQWinStd(CalcId, SampleId, dbPath = "X:/EQWin/WR/DB/Water Resources.mdb")
}
\arguments{
\item{CalcId}{The ID identifying the calculation to perform, from column CalcId of table eqcalcs.}

\item{SampleId}{The sample ID for the calculation, from column SampleId of table eqsampls. This is required to retrieve the additional parameter values and the sample date needed for the calculation.}

\item{dbPath}{The path to the EQWin database. Default is "X:/EQWin/WR/DB/Water Resources.mdb".}
}
\value{
A value for the calculated standard.
}
\description{
The EQWin database stores formulas for calculating standards. This function retrieves the formula, parses it into R code, and calculates the standard for a specified parameter. As these calculations typically require additional measured values, the SampleId is also required. Note that the target parameter is NOT required; this is identified by the CalcId.
}
