% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFlowStats.R
\name{estimateFlowStats}
\alias{estimateFlowStats}
\title{Function that predicts water flow statistics at creeks}
\usage{
estimateFlowStats(
  gauged_stations,
  ungauged_area,
  perc = 95,
  record_length = 15
)
}
\arguments{
\item{gauged_stations}{The stations used for estimating flow. Given as a vector of the staton code. Ex: c(08AA008, 09BC001).}

\item{ungauged_area}{The area of the stations of interest. Given in square km.}

\item{perc}{The percent of data (days) required to include a year or season in the statistics. Default is 95.}

\item{record_length}{The length of record (# of years/seasons) at a station required to be included in the statistics. Default is 15.}
}
\value{
A table and a plot in a list. The table contains the estimated stats for the river/stream/creek of interest. The table contains 3 columns: season, stat and value. The season column refers to the season for which the stat was calculated, where annual is the entire year, fall is Sept. 1st - Oct 31st, freshet is April 20 - June 30, summer is July 1st - August 31st, and winter is Nov. 1st - April 19. The stat column column refers to the type of statistic. The difference between max and max_all is that max is the average maximum for that season and the max_all is the all-time maximum for that season.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The purpose of this script is to estimate flows of rivers/streams/creeks based on the flow at multiple other stations of choice. It was created because Highways (YG) often wants flows statistics for small streams from which water is pulled during highway work.

Statistics for each station are averaged over all years/seasons available, except for all-time data points. Years or seasons with more than 5 percent missing data (this is the default) are excluded from the statistics. Additionally, only stations with a minimum of 15 years/seasons of data (this is the default) are retained. Next, statistics are expressed as m3/s/km2 and plotted to visualize outliers (plot). Each statistic is averaged for all stations and applied to the ungauged rivers/streams/creeks using their catchment area.
}
