% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshetReport.R
\name{freshetReport}
\alias{freshetReport}
\title{Freshet condition reporting utility - public}
\usage{
freshetReport(
  report_name = "Territory",
  custom_report_stations = NULL,
  extra_years = NULL,
  preset_extra_years = FALSE,
  report_type = "Level",
  plot_titles = FALSE,
  level_zoom = TRUE,
  flow_zoom = NULL,
  zoom_days = 20,
  rate = TRUE,
  level_returns = "table",
  flow_returns = "table",
  MESH = FALSE,
  force_CGVD28 = FALSE,
  CLEVER = FALSE,
  precip = FALSE,
  meteogram = FALSE,
  WSC_images = FALSE,
  image_path = NULL,
  save_path = "choose"
)
}
\arguments{
\item{report_name}{The name of the report you wish to generate. One of "Dawson", "Southern Lakes", "Carmacks", "Ross/Pelly", "Mayo/Stewart", "Liard/Watson Lake", "Teslin", "Old Crow", "Aishihik", "Alsek", "Territory" (for an overview of the territory with fewer stations). Most minor spelling variations should work. Defaults to "Territory".}

\item{custom_report_stations}{A user-specified list of stations for which to generate a report. Defaults to NULL to operate on the report_name parameter instead. Input must be a character vector of station IDs, as in c("station1", "station2"). Reminder: you can create a character vector from a column of a data.frame, and you can reference an environment object instead of typing in the vector!}

\item{extra_years}{Specify extra years of data to plot for one or multiple stations. Use the form "09AB001:1990,2020", "09EB003:1985". Concatenate if more than one station. Can be used together with preset_extra_years}

\item{preset_extra_years}{TRUE or FALSE, defaults to FALSE. Can be used together with extra_years.}

\item{report_type}{What do you want your report to contain? Choose from "Level", "Flow", or "Both." Defaults to Both.}

\item{plot_titles}{Do you want the plots to have a title?}

\item{level_zoom}{Do you want a zoomed-in plot for level? Choose from TRUE or FALSE. Defaults to TRUE.}

\item{flow_zoom}{Do you want a zoomed-in plot for flow? TRUE/FALSE/NULL, defaults to NULL which copies the setting for level_zoom.}

\item{zoom_days}{The number of days to plot for zoomed in level plots. Defaults to 30, but not used unless level_zoom is set to TRUE.}

\item{rate}{Should rates of change for flow and level be included? The 24-hour rate of change will be plotted on zoomed-in graphs on the right y-axis, and a rate table will be included. TRUE/FALSE, defaults to TRUE.}

\item{level_returns}{Should level returns be calculated, plotted, and added to the level table? You have the option of using pre-determined level returns only (option "table"), auto-calculated values with no human verification (option "calculated", calculated on-the-fly using all data available from March to September, up to the current date), "auto" (priority to pre-determined levels, fallback to auto-calculated levels), or none (option "none"). Defaults to "auto".}

\item{flow_returns}{Should flow returns be calculated, plotted, and added to the flows table? You have the option of using pre-determined flow returns only (option "table"), auto-calculated values with no human verification (option "auto", calculated on-the-fly using all data available from March to September, up to the current date), "both" (with priority to pre-determined flows), or none (option "none"). Defaults to "both".}

\item{MESH}{Should MESH forecasts be incorporated into the graphs?}

\item{force_CGVD28}{For stations with a datum, should CGVD28 be used even if there is a more recent datum?}

\item{CLEVER}{Should CLEVER forecasts be incorporated into the graphs?}

\item{precip}{Should precipitation data (accumulated precip above stations) and images (precip across whole territory) be included? TRUE or FALSE}

\item{meteogram}{Should meteograms relevant to the stations in the report be included? TRUE or FALSE.}

\item{WSC_images}{Should images from WSC fixed cameras be included? TRUE or FALSE.}

\item{image_path}{The path to the directory (folder) containing the images you wish to include. Default to NULL to not include any extra images. Set to "choose" to navigate to the folder, or enter the folder path directly as a character string. Some reports automatically include web-hosted images, do not include them here. WARNING: option 'choose' only works on Windows, and some late-build R versions have a bug that prevents it from working every time.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default "choose" lets you select your folder, otherwise enter the path as a character string. WARNING: option 'choose' only works on Windows, and some late-build R versions have a bug that prevents it from working every time.}
}
\value{
A flood report containing flow and water level information in Microsoft Word format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function generates condition reports for preset or user-specified Water Survey of Canada stations, in a format intended for public consumption. In addition to water level and flow, precipitation data, still images, and weather forecasts are incorporated. The output is a Microsoft Word document on a Yukon Government template.

This function fetches data directly from the Water Survey of Canada and does not use the local hydrometric database created/maintained by the AquaCache package.

Parts of this report fetch password-protected information:
To download real-time WSC data, you MUST have your hydat credentials loaded into your .Renviron profile as values pairs of WS_USRNM=”your_username” and WS_PWD=”your_password”.
To download WSC images, you MUST have your ECCC credentials loaded into your .Renviron profile as value pairs of ECCCUSER="your_username" and ECCCPASS="your_password".  Refer to the R and GitHub for the WRB word document for more information.
}
