% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{get_bulletin_value}
\alias{get_bulletin_value}
\title{Get bulletin value for each station}
\usage{
get_bulletin_value(bulletin_month, bulletin_year, ts, parameter)
}
\arguments{
\item{bulletin_month}{Integer bulletin month}

\item{bulletin_year}{Integer bulletin year}

\item{ts}{Wide-format data.frame with 'datetime' column and station columns}

\item{parameter}{Character string specifying the parameter name.}
}
\value{
Named numeric vector of bulletin values for each station
}
\description{
Get bulletin value for each station
}
