% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cached}
\alias{get_cached}
\title{In-memory cache}
\usage{
get_cached(key, fetch_fun, check_fun = NULL, ttl = 3600, env = .GlobalEnv)
}
\arguments{
\item{key}{A unique character string identifying the cached data.}

\item{fetch_fun}{A function that returns the value to cache when the current cache entry is missing or expired.}

\item{check_fun}{An optional function that checks if the cached value is still valid even though the \code{ttl} is expired and returns TRUE or FALSE. If TRUE, the cached value is returned; if FALSE, a new value is fetched. This function should be significantly faster than \code{fetch_fun}.}

\item{ttl}{Time-to-live for the cached value in seconds. Defaults to 3600 seconds (1 hour).}

\item{env}{The environment in which to store the cache. Defaults to \code{.GlobalEnv}, but can be set to a different environment if needed, such as session$userData in a Shiny app so that the cache is specific to that session.}
}
\value{
The cached value for \code{key}.
}
\description{
Provides global helper functions to store and retrieve cached data across user sessions. Each cached item includes a timestamp and is automatically refreshed when called after the specified time-to-live (TTL) has expired.
}
