% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{get_future_flow_conditions}
\alias{get_future_flow_conditions}
\title{Get flow conditions for stations based on current water level and flow compared to thresholds}
\usage{
get_future_flow_conditions(hydrometric, swe, language = "English")
}
\arguments{
\item{hydrometric}{scalar flow 'relative_to_med' value}

\item{swe}{scalar swe 'relative_to_med' value}

\item{language}{Character string indicating the language for labels and legends (default "English")}
}
\description{
Get flow conditions for stations based on current water level and flow compared to thresholds
}
