% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{get_period_dates}
\alias{get_period_dates}
\title{Helper: get start/end dates for a given parameter and year/month}
\usage{
get_period_dates(year, month, october_start = FALSE)
}
\arguments{
\item{year}{bulletin year}

\item{month}{bulletin month}

\item{october_start}{Norm period starts in October of previous year regardless of bulletin month}
}
\value{
start date and end date for norm calculation period
}
\description{
Helper: get start/end dates for a given parameter and year/month
}
