% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getDocument}
\alias{getDocument}
\title{Extract document from postgreSQL DB}
\usage{
getDocument(id, con = NULL, save_dir = NULL, save_name = NULL)
}
\arguments{
\item{id}{The ID number from column 'document_id' of table 'documents'.}

\item{con}{A connection to the database. NULL will use \code{\link[=AquaConnect]{AquaConnect()}} and disconnect automatically when done.}

\item{save_dir}{A directory in which to write the file.}

\item{save_name}{The name to give the file, \emph{without} extension.}
}
\value{
A data.frame containing the row(s) identified by the specified ID. One of the columns will contain the binary object (file). The document will be saved to file if requested and possible.
}
\description{
lifecycle::badge("stable")
Extracts a document stored in the aquacache as BYTEA type. For extracting images see \code{\link[=getImage]{getImage()}}, for extracting vector files as \emph{terra} rasters (points, lines polygons) see \code{\link[=getVector]{getVector()}}, and for extracting rasters to R as \emph{terra} objects see \code{\link[=getRaster]{getRaster()}}.
}
\details{
If you need additional flexibility use function \code{\link[=getFile]{getFile()}} instead. This should not normally be needed with a database created by the AquaCache package but is made available for use with other databases.
}
