% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFile}
\alias{getFile}
\title{Extract document stored as BYTEA from postgreSQL DB}
\usage{
getFile(id, id_col, ext, table, con = NULL, save_dir = NULL, save_name = NULL)
}
\arguments{
\item{id}{The unique ID from the column specified in \code{id_col} with which to identify the individual record containing the binary object.}

\item{id_col}{The column in which to look for the \code{id}}

\item{ext}{The column in which to look for the file extension, or a file extension (must be preceded by a period). Only used if save_path is not NULL}

\item{table}{The table to look in for the document.}

\item{con}{A connection to the database. NULL will use \code{\link[=AquaConnect]{AquaConnect()}} and disconnect automatically when done.}

\item{save_dir}{A directory in which to write the file.}

\item{save_name}{The name to give the file, \emph{without} extension (specify the extension in \code{ext}).}
}
\value{
A data.frame containing the row(s) identified by the specified ID. One of the columns will contain the binary object (file). The document will be saved to file if requested and possible.
}
\description{
\emph{NOTE} Unless you really need the additional flexibility use functions \code{\link[=getDocument]{getDocument()}} or \code{\link[=getImage]{getImage()}}.
lifecycle::badge("stable")

Extracts a document/file stored in a postgreSQL database as BYTEA type. Should be flexible enough to work with most schemas, but was designed around the 'documents' and 'images' tables created with this package and with files uploaded to the database with functions from package AquaCache. For extracting vector files as \emph{terra} rasters (points, lines polygons) see \code{\link[=getVector]{getVector()}}, and for extracting rasters to R as \emph{terra} objects see \code{\link[=getRaster]{getRaster()}}.
}
