% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHRDPA.R
\name{getHRDPA}
\alias{getHRDPA}
\title{Download/process ECCC HRDPA rasters}
\usage{
getHRDPA(
  start = Sys.time() - 60 * 60 * 24,
  end = Sys.time(),
  clip = c("YT"),
  save_path = "choose"
)
}
\arguments{
\item{start}{The start of the time period for which you want HRDPA rasters, up to one month prior to today. Use format \code{"yyyy-mm-dd hh:mm"} (character vector) in local time. See details for more info.}

\item{end}{The end of the time period for which you want HRDPA rasters. Other details as per \code{start}.}

\item{clip}{The two-digit abbreviation(s) as per \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=t1_8}{Canadian Census} for the province(s) with which to clip the HRDPA. A 300 km buffer is added beyond the provincial boundaries. Set to NULL for no clip.}

\item{save_path}{The path to the directory (folder) where the raster(s) should be saved. Default \code{"choose"} lets you select your folder, or enter the path as a character string. Any necessary rasters already in the save path will not be downloaded again.}
}
\value{
One of more rasters of precipitation amounts over 6 hour periods saved where specified. The date/time stamp in the file name refers to the end of the 6-hour valid period. Output is not assigned to an object, load the raster from the save_path if desired.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Utility function to retrieve and minimally process precipitation reanalysis rasters generated by the Canadian Precipitation Analysis (CaPA) System \href{https://eccc-msc.github.io/open-data/msc-data/nwp_hrdpa/readme_hrdpa_en/}{HRDPA}, clips them to a designated province/region, and saves them to a designated location in .tiff format.

CaPA outputs precipitation estimates valid at  00, 06, 12 and 18 UTC for the 6-hour rasters. The start and end times you specify will be automatically adjusted to the nearest or most recent match. Note that there is a 1-hour delay in the issuing of CaPA outputs. In addition, a preliminary version of the output is issued ~one hour post with a final version output ~7 hours post incorporating observation data for model corrections. This function will look for 1-hour post rasters in the save path and replace them with 7-hour post files whenever possible.

Start and end times are automatically converted to UTC if supplied as class "character"; input times/dates of class POSIXct if you wish to specify a time zone.

Be aware that your requested time period may not be realized when requesting an end time close to the current time. This will occur if ECCC experiences an unusual delay in the issuance of the 1-hour raster. These are normally issued prior to 1 hour and 10 minutes of the valid time, issues may arise is they are delayed by more than 1 hour and 12 minutes.
}
\seealso{
\code{\link[=getHRDPS]{getHRDPS()}} if looking for forecast precipitation and other parameters in .tif format. For nice precipitation maps and tabular reports of precipitation (past or future), try \code{\link[=basinPrecip]{basinPrecip()}}.
}
