% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatial.R
\name{getRaster}
\alias{getRaster}
\title{Retrieve raster files from the database}
\usage{
getRaster()
}
\value{
An error message telling the user what to do instead.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Placeholder function, not yet complete.
}
