% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatial.R
\name{getRasters}
\alias{getRasters}
\title{Retrieve raster files from the database}
\usage{
getRasters(
  raster_reference_ids = NULL,
  boundary = NULL,
  bands = 1,
  tbl_name = c("spatial", "rasters"),
  col_name = "rast",
  con = NULL,
  stack = TRUE
)
}
\arguments{
\item{raster_reference_ids}{A vector of raster reference IDs to filter the raster data and return the results. For example, c(1, 2, 3). If NULL, all raster data is returned.}

\item{boundary}{Optional spatial boundary to limit the extent of the raster data returned, as a terra::spatVector object or a numeric vector c(top, bottom, right, left) indicating the projection-specific limits with which to clip the raste (by default, decimal degree lat/long).  If NULL, the full extent of the raster data is returned.}

\item{bands}{The raster bands to return. Default is 1, which returns the first band of the raster data. If you want to return all bands, set this to TRUE.}

\item{tbl_name}{The name of the schema and table containing the raster data. Default is c("spatial", "rasters").}

\item{col_name}{The name of the column containing the raster data. Default is "rast".}

\item{con}{A connection to the target database. If NULL, a new connection is created using \code{\link[=AquaConnect]{AquaConnect()}} and automatically closed when the function exits.}

\item{stack}{Logical indicating whether to stack the retrieved rasters into a single terra::rast object. Default is TRUE.}
}
\value{
A terra::rast or list of terra::rast objects containing the raster data.
}
\description{
This function retrieves raster data from a PostGIS-enabled database and returns it as a terra object. It allows for filtering of the raster data using SQL clauses, spatial boundaries, and selection of specific raster bands.
}
