% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeather.R
\name{getWeather}
\alias{getWeather}
\title{Download ECCC weather station data}
\usage{
getWeather(
  station,
  start,
  end = Sys.Date(),
  tzone = "UTC",
  interval = "hour",
  save_path = NULL
)
}
\arguments{
\item{station}{The station for which you want data. You can specify the 7-digit/letter Climate ID, the 4 or 5 digit ECCC station ID, the 5-digit WMO ID (starts with a 7), or the three-letter Transport Canada ID (i.e YDA and not CYDA). If working interactively you can also specify the station name or part thereof (as character vector) and select from a list.}

\item{start}{The start date for which you want data. Input either a character vector of form "2022-12-30" or a Date formatted object.}

\item{end}{The end date for which you want data. Input either a character vector of form "2022-12-30" or a Date formatted object.}

\item{tzone}{Choose from "local" or "UTC". Note that "local" is whatever ECCC thinks the station local TZ is: this might be inaccurate in Yukon, where many (maybe even all) stations report in UTC-8.}

\item{interval}{Select from 'hour', 'day', or 'month'. However, if requesting a location by name you will need to select your chosen interval  interactively (so don't bother modifying the default).}

\item{save_path}{The path where you wish to save the resultant .csv file. Defaults to NULL, in which case you should assign the function to an R object. Set to "choose" to interactively select the location.}
}
\value{
A data.frame of weather data and, if save_path is specified, a csv of this same data located in the save_path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This script downloads data from an ECCC station for a given date range, calling \code{\link[weathercan:weather_dl]{weathercan::weather_dl()}} to download the data. This function facilitates interaction with that package by modifying start and end dates if your request is out of range, and allows you to interactively search for locations by name. Note that this function may take a long time to complete if you are requesting multiple years of data!
}
\seealso{
\code{\link[weathercan:weather_dl]{weathercan::weather_dl()}} for a simpler, pared-down means of downloading ECCC weather data. For nice precipitation maps and tabular reports of precipitation (past or future), try \code{\link[=basinPrecip]{basinPrecip()}}.
}
