% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_processing.R
\name{html_processing}
\alias{html_processing}
\title{In Situ html logger file processing}
\usage{
html_processing(
  html_file,
  master_sheet =
    "G:/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/1_YOWN_MASTER_TABLE/YOWN_MASTER.xlsx",
  logger_tracking =
    "G:/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/3_OTHER/YOWN_Logger_Tracking.xlsx",
  YOWN_logger_folder =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/9_LOGGER_FILE_DROPBOX"
)
}
\arguments{
\item{html_file}{File name of .html file, must be in the logger dropbox folder}

\item{master_sheet}{Path to YOWN master sheet}

\item{logger_tracking}{Path to YOWN logger tracking sheet}

\item{YOWN_logger_folder}{Path to YOWN logger dropbox folder}
}
\value{
Moves YOWN html file to backups folder and appropriate YOWN Active Wells folder. Uploads data to Aquarius after performing unit checks and conversions. Adds pressure column if missing, based on depth column and conversions calcs on In Situ website
}
\description{
# Parse YOWN html logger files, upload to Aquarius, sort into respective folder, and track logger metadata. Used "Pressure" column after converting to m water column. If pressure column does not exist, converts from deth using formula supplied by In Situ
}
