% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_check.R
\name{hydat_check}
\alias{hydat_check}
\title{Check and update HYDAT}
\usage{
hydat_check(silent = FALSE, path = NULL)
}
\arguments{
\item{silent}{Should messages be output to the console?}

\item{path}{The path to the folder where the HYDAT database is stored or to be stored, or the path to the Hydat.sqlite3 file itself. If NULL, the function will attempt to find the database in the default location. Don't use this parameter unless you have a good reason to.}
}
\value{
An updated local copy of hydat, if indicated by the age of the remote HYDAT.
}
\description{
Checks and, if necessary, updates the local version of the WSC HYDAT database. Intended for use on a schedule to ensure the user always has the latest version.
}
