% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{hydrometConnect}
\alias{hydrometConnect}
\title{Connect to the hydromet database}
\usage{
hydrometConnect(
  name = "hydromet",
  host = Sys.getenv("hydrometHost"),
  port = Sys.getenv("hydrometPort"),
  username = Sys.getenv("hydrometUser"),
  password = Sys.getenv("hydrometPass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form hydrometHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter=value pair of form hydrometPort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter=value pair of form hydrometUser="username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter=value pair of form hydrometPass="password".}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Superseded by function \code{\link[=AquaConnect]{AquaConnect()}}. Still exported by this package for backwards compatibility, as old database 'hydromet' is now a development environment.

This function exists to facilitate connecting to the hydrology database. Database superusers will have access to all database records, but other users will be asked to provide their username and password to access records other than the 'public' ones. Note that this is \emph{not necessarily} the same username and password as the one used to log into the database itself.
}
