% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrometDiscrete.R
\name{hydrometDiscrete}
\alias{hydrometDiscrete}
\title{Discrete hydrometric data plotting}
\usage{
hydrometDiscrete(
  location = NULL,
  parameter,
  startDay = 1,
  endDay = 365,
  binwidth = "1 month",
  tzone = "MST",
  years = NULL,
  title = TRUE,
  custom_title = NULL,
  plot_type = "violin",
  plot_scale = 1,
  save_path = NULL,
  con = NULL,
  discrete_data = NULL,
  ref_period_start_datetime = NULL,
  ref_period_end_datetime = NULL,
  lang = "en"
)
}
\arguments{
\item{location}{The location for which you want a plot. Can be left NULL if \code{discrete_data} is provided.}

\item{parameter}{The parameter you wish to plot. The location:parameter combo must be in the local database unless \code{discrete_data} is provided. In this later case, the parameter must still be specified to label the y-axis.}

\item{startDay}{Can be specified as a number from 1 to 365, as a character string of form "yyyy-mm-dd", or as a date object. Either way the day of year is the only portion used, specify years to plot under parameter \code{years}.}

\item{endDay}{The end day of year for the plot x-axis. As per \code{startDay}.}

\item{binwidth}{The width of the bins in the violin/boxplot in days, specified as an integer of \strong{n} days or as "1 day", "1 week", "1 month", etc. Default is one month.}

\item{tzone}{The timezone to use for fetching data. Datetimes are stored in the database in UTC offset, so this parameter could make a difference to what day a particular sample is considered to be on. In most cases you can ignore this parameter.}

\item{years}{The years to plot. If \code{startDay} and \code{endDay} cover December 31 - January 1, select the December year(s). Max 10 years, NULL = current year.}

\item{title}{Should a title be included?}

\item{custom_title}{Custom title to be given to the plot. Default is NULL, which will set the title as such: Location 09AB004: Marsh Lake Near Whitehorse.}

\item{plot_type}{Choose from "violin" , "boxplot" or "linedbox".}

\item{plot_scale}{Adjusts/scales the size of plot text elements. 1 = standard size, 0.5 = half size, 2 = double the size, etc. Standard size works well in a typical RStudio environment.}

\item{save_path}{Default is NULL and the graph will be visible in RStudio and can be assigned to an object. Option "choose" brings up the File Explorer for you to choose where to save the file, or you can also specify a save path directly.}

\item{con}{A connection to the target database. NULL uses \code{\link[=AquaConnect]{AquaConnect()}} and automatically disconnects.}

\item{discrete_data}{A data.frame with the data to be plotted. Must contain the following columns: datetime, year, month, value and units.}

\item{ref_period_start_datetime}{The first year to consider when calculating ribbons for a reference period. Default is NULL and will use the first year of data available.}

\item{ref_period_end_datetime}{The last year to consider when calculating ribbons for a reference period. Default is NULL and will use the last year of data available or the last}

\item{lang}{The desired language for the plot.}
}
\value{
A .png file of the plot requested (if a save path has been selected), plus the plot displayed in RStudio. Assign the function to a variable to also get a plot in your global environment as a ggplot object which can be further modified
}
\description{
Deprecated for general usage but retained for use with snow bulletin. Use \code{\link[=plotDiscrete]{plotDiscrete()}} instead.

Generate plots of snow survey data (snow water equivalent and snow depth) or other variables sampled at regular intervals (weekly or monthly).

Notice: in many cases, you're better off using the Shiny app at \code{\link[=AquaLink]{AquaLink()}} to generate and export your plot. Read on if you need additional control over the final product.

This function plots data from the aquacache database (maintained by the AquaCache package) and yields consistent-looking plots for discrete data. This function can only plot what's in the database. Data can be represented as violin plots, as regular box plots or as a 'linedbox' plot (imitates plots currently used in the snow bulletin).
}
