% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnR.R
\name{learnR_answer}
\alias{learnR_answer}
\title{Answers for the learnR exercises}
\usage{
learnR_answer(exercise, part = 1)
}
\arguments{
\item{exercise}{The exercise number for which you want the answer.}
}
\value{
The answer for the specified exercise, both as text and as a command to be executed in the console.
}
\description{
Answers for the learnR exercises
}
\examples{

learnR_answer(1)
}
