% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnR.R
\name{learnR_hint}
\alias{learnR_hint}
\title{Hints for the learnR exercises}
\usage{
learnR_hint(exercise)
}
\arguments{
\item{exercise}{The exercise number for which you want a hint.}
}
\value{
A hint for the specified exercise.
}
\description{
Hints for the learnR exercises
}
\examples{

learnR_hint(1)
}
