% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\docType{data}
\name{location}
\alias{location}
\title{Plot a continous timeseries from the hydromet database}
\format{
An object of class \code{character} of length 1.
}
\usage{
location
}
\arguments{
\item{location}{The location for which you want a plot.}

\item{parameter}{The parameter you wish to plot. The location:parameter combo must be in the local database.}

\item{record_rate}{The recording rate for the parameter and location to plot. In most cases there are not multiple recording rates for a location and parameter combo and you can leave this NULL. Otherwise NULL will default to the most frequent record rate, or set this as one of '< 1 day', '1 day', '1 week', '4 weeks', '1 month', 'year'.}

\item{start_date}{The day on which to start the plot as character, Date, or POSIXct. Default is 1970-01-01 but the plot is zoomable.}

\item{end_date}{The day on which to end the plot as character, Date, or POSIXct.. Default is today.}

\item{title}{Should a title be included?}

\item{custom_title}{Custom title to be given to the plot. Default is NULL, which will set the title as Location <\if{html}{\out{<location id>}}>: <\if{html}{\out{<location name>}}>. Ex: Location 09AB004: Marsh Lake Near Whitehorse.}

\item{filter}{Should an attempt be made to filter out spurious data? Will calculate the rolling IQR and filter out clearly spurious values. Set this parameter to an integer, which specifies the rolling IQR 'window'. The greater the window, the more effective the filter but at the risk of filtering out real data. Negative values are always filtered from parameters "level", "flow", "snow depth", "SWE", "distance", and any "precip" related parameter. Otherwise all values below -100 are removed.}

\item{historic_range}{Should the historic range be plotted? Default is TRUE.}

\item{returns}{Should returns be plotted? You have the option of using pre-determined level returns only (option "table"), auto-calculated values (option "calculate"), "auto" (priority to "table", fallback to "calculate"), or "none". Defaults to "auto".}

\item{return_type}{Use minimum ("min") or maximum ("max") values for returns?}

\item{return_months}{Numeric vector of months during which to look for minimum or maximum values. Only works with calculated returns. Does not have to be within \code{startDay} and \code{endDay}, but will only consider data up to the last year specified in \code{years}. For months overlapping the new year like November-April, should look like c(11:12,1:4). IMPORTANT: the first month in the range should be the first element of the vector: c(1:4, 11:12) would not be acceptable. Think of it as defining a season. Passed to 'months' argument of \code{\link[fasstr:calc_annual_extremes]{fasstr::calc_annual_extremes()}} and also used to set the 'water_year_start' parameter of this function.}

\item{return_max_year}{The last year of data to consider when calculating returns. If left NULL behavior depends on parameter \code{historic_range}: if \code{historic_range} is set to 'last' defaults to the last year in \code{year} otherwise uses all data available. Automatically set to max(years) if \code{historic_range} is 'last', otherwise set to the current year.}

\item{allowed_missing}{Allowable \% of data missing during the months specified in 'return_months' to still retain the year for analysis when calculating returns. Passed to 'allowed_missing' argument of \code{\link[fasstr:calc_annual_extremes]{fasstr::calc_annual_extremes()}}.}

\item{plot_scale}{Adjusts/scales the size of plot text elements. 1 = standard size, 0.5 = half size, 2 = double the size, etc. Standard size works well in a typical RStudio environment.}

\item{legend}{Should a legend (including text for min/max range and return periods) be added to the plot?}

\item{save_path}{Default is NULL and the graph will be visible in RStudio and can be assigned to an object. Option "choose" brings up the File Explorer for you to choose where to save the file, or you can also specify a save path directly.}

\item{con}{A connection to the database. Default uses function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\description{
Plot a continous timeseries from the hydromet database
}
\keyword{datasets}
