% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_maputils.R
\name{remove_unused_bins}
\alias{remove_unused_bins}
\title{Remove unused bins from dynamic style elements based on the presence of data in point datasets}
\usage{
remove_unused_bins(
  dynamic_style_elements,
  point_data,
  point_data_secondary,
  statistic
)
}
\arguments{
\item{dynamic_style_elements}{List containing bins, colors, and labels for styling}

\item{point_data}{sf object containing point-based station data (surveys)}

\item{point_data_secondary}{sf object containing secondary point data (pillows)}

\item{statistic}{Character string indicating which SWE value to visualize}
}
\value{
Updated dynamic_style_elements with unused bins removed if necessary
}
\description{
For the "relative_to_med" statistic, checks if the -2 and -1 bins are present in the point datasets. If not, removes those bins from the dynamic style elements to prevent them from appearing in the legend.
}
