% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shortenLanguage}
\alias{shortenLanguage}
\title{Standardize Language Input to short form}
\usage{
shortenLanguage(language)
}
\arguments{
\item{language}{Character string specifying the language. Accepts values such as
"en", "english", "anglais" for English, and "fr", "french", "français", "francais",
"français" for French. Case-insensitive.}
}
\value{
A character string: "en" for English or "fr" for French.
}
\description{
Standardizes the input language to either "en" (English) or "fr" (French).
Accepts various representations of English and French, including common abbreviations
and translations. If the input is \code{NULL}, defaults to English. If the input is not
recognized, a warning is issued and the function defaults to English.
}
\examples{
shortenLanguage("EN")        # returns "en"
shortenLanguage("français")  # returns "fr"
shortenLanguage(NULL)        # returns "en"
shortenLanguage("de")        # returns "en" with a warning
}
