% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowBulletin.R
\name{snowBulletin}
\alias{snowBulletin}
\title{Function for creating snow bulletin (Rmarkdown)- internal}
\usage{
snowBulletin(sub_basin_name = NULL, year = NULL, save_path = "choose")
}
\arguments{
\item{sub_basin_name}{The name of the sub_basin you wish to generate. One or many of "Upper Yukon", "Teslin", "Central Yukon", "Pelly", "Stewart", "White", "Lower Yukon", "Porcupine", "Peel", "Liard", "Alsek". North Slope will be added when hydromet is updated with the new snow database. Default is NULL, where all basins are shown in bulletin.}

\item{year}{Year for which the snow bulletin is to be created.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default "choose" lets you select your folder, otherwise enter the path as a character string. WARNING: option 'choose' only works on Windows, and some late-build R versions have a bug that prevents it from working every time.}
}
\value{
A snow bulletin in Microsoft Word format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function generates the snow bulletin for a specified date. The output is a Microsoft Word document on a Yukon Government template.

This function fetches data from the local postgresql hydrometric database created/maintained by the HydroMetDB package.
}
\details{
To download data, you MUST have your hydromet credentials loaded
into your .Renviron profile as values pairs of hydrometHost="10.250.12.154", hydrometPort="5433", hydrometUser="hydromet_read", hydrometPass="hydromet".
}
