% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowBulletinStats.R
\name{snowBulletinStats}
\alias{snowBulletinStats}
\title{Function for creating snow bulletin statistics - internal}
\usage{
snowBulletinStats(
  year,
  month,
  basins = NULL,
  excel_output = FALSE,
  save_path = "choose",
  synchronize = FALSE,
  source = "AquaCache"
)
}
\arguments{
\item{year}{Year for which the snow bulletin stats are calculated.}

\item{month}{Month for which the snow bulletin stats are calculated. Options are 3, 4 or 5.}

\item{basins}{The name of the sub_basin you wish to generate stats for. One or many of "Upper Yukon", "Teslin", "Central Yukon", "Pelly", "Stewart", "White", "Lower Yukon", "Porcupine", "Peel", "Liard", "Alsek". North Slope will be added when AquaCache is updated with the new snow database. Default is NULL, where all basins are shown in bulletin.}

\item{excel_output}{If TRUE, calculated stats will be outputted in multiple excel tables.}

\item{save_path}{The path to the directory (folder) where the excel files should be saved. Enter the path as a character string.}

\item{synchronize}{Should the timeseries be synchronized with source data? If TRUE, all timeseries used in the snow bulletin will be synchronized. If FALSE (default), none will be synchronized. Currently is a placeholder and does not work.}

\item{source}{Should the SWE statistics for stations (station_stats) be calculated from the hydro database or the snow database. Default is hydro database.}
}
\value{
A snow bulletin in Microsoft Word format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function generates the statistics associated with the snow bulletin for a specified date. The output is list of tables (in R), or a number of excel workbooks.

This function fetches data from the local postgresql AquaCache database created/maintained by the AquaCache package.
}
\details{
To download data, you MUST have your AquaCache credentials added to your .Renviron file. See function \code{\link[=AquaConnect]{AquaConnect()}} or talk to the database administrator/data scientist for help.
}
