% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowConnect_access.R
\name{snowConnect_access}
\alias{snowConnect_access}
\title{Connect to snow survey database}
\usage{
snowConnect_access(path = "default", silent = FALSE)
}
\arguments{
\item{path}{The path to the database. Currently supports either a Microsoft Access database or an .sqlite database. 'default' points to //carver/infosys/Snow/DB/SnowDB.mdb.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Connects to the Snow Survey Access database, or, if this DB is supersceded by a .sqlite database, to an equivalent .sqlite DB. If the Snow Survey database is migrated to an SQLite database, beware that the tables and table columns might no longer be what you're expecting!
}
