% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_basin.R
\name{SWE_basin}
\alias{SWE_basin}
\title{Calculates snow survey stats for basins}
\usage{
SWE_basin(
  year,
  month,
  threshold = 7,
  csv = FALSE,
  summarise = FALSE,
  source = "aquacache",
  con = NULL
)
}
\arguments{
\item{year}{The year of interest. If summarise = TRUE, the stats will be calculated based on all years prior to 'year'. If summarise = FALSE, only data from the current year and before are taken.}

\item{month}{The month of interest. Options are 3, 4 and 5 for March, April and May, respectively. Can also give multiple months as a vector. Historical stats are given for the first day of this month.}

\item{threshold}{A number between 1 and 10 giving the threshold below which the SWE for that basin and year are ignored. These numbers represent the sum of the factors of the stations for a basin which are not missing data for that year. 10 means that the swe values calculated from less than all the stations of that basin are ignored. 1 means that only the swe calculated from less than 1 out of 10 are ignored.}

\item{csv}{TRUE or FALSE. If TRUE, a csv will be created.}

\item{summarise}{TRUE to summarise the data into a data.frame with the current SWE, historical median, the swe relative to the median (swe / swe_median), historical maximum, historical minimum, and year of maximum and minimum for each basin.}

\item{source}{Database from which to fetch this data. Options are: aquacache or snow.}

\item{con}{A connection to the database specified in \code{source}. If NULL, a new connection will be created with \code{\link[=AquaConnect]{AquaConnect()}} or \code{\link[=snowConnect]{snowConnect()}} (depending on 'source' parameter) and automatically closed.}
}
\value{
A table and a csv file (if csv = TRUE) with either (summarise = FALSE) the swe for all basins, years and months of interest or (summarise = TRUE) the current SWE, historical median, the swe relative to the median (swe / swe_median), historical maximum, historical minimum, and year of maximum and minimum for each basin and month of interest.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
The purpose of this script is to summarise the SWE data of each basin for a particular year and month and compare to previous years. It is used for the snow bulletin, specifically the SWE map and the plot B. It is meant to replace Ellen Ward's code from 2020-04-16, r script called swe_compiled_basin.R.
}
