% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleCase.R
\name{titleCase}
\alias{titleCase}
\title{Capitalize titles in French or English}
\usage{
titleCase(text, language = "en")
}
\arguments{
\item{text}{The text to capitalize.}

\item{language}{The language to use for capitalization, limited to 'fr' or 'en'.}
}
\value{
A string with proper capitalization for the language.
}
\description{
Capitalize titles in French or English
}
