% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tr}
\alias{tr}
\title{Find the translation for a given placeholder}
\usage{
tr(key, lang, translations)
}
\arguments{
\item{key}{A string representing the translation key. Must be a name in the 'id' column of the translations data.frame, else an error will be generated.}

\item{lang}{A string representing the target language code. Must be a name in the translations list, else an error will be generated.}

\item{translations}{A named list (with names representing the 'lang') of named character vectors (with names as the 'key') containing translations for different languages.}
}
\description{
This function retrieves the translation for a specified key and language
from a provided translations list. Allows for referencing only the key in
function code, with the language specified at runtime.
}
