% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQreport.R
\name{WQreport}
\alias{WQreport}
\title{Create a tabular report of water quality data from the EQWin database}
\usage{
WQreport(
  date,
  stations = NULL,
  stnGrp = NULL,
  parameters = NULL,
  paramGrp = NULL,
  stds = NULL,
  stnStds = TRUE,
  date_approx = 0,
  save_path = "choose",
  dbPath = "X:/EQWin/WR/DB/Water Resources.mdb"
)
}
\arguments{
\item{date}{The date for which to fetch results as a Date object or character vector that can be coerced to a date.}

\item{stations}{A vector of station names as listed in the EQWiN eqstns table, column StnCode. Leave NULL to use stnGrp instead.}

\item{stnGrp}{A station group as listed in the EWQin eqgroups table, column groupname. Leave NULL to use stations instead.}

\item{parameters}{A vector of parameter names as listed in the EQWin eqparams table, column ParamCode. Leave NULL to use paramGrp instead.}

\item{paramGrp}{A parameter group as listed in the EQWin eqgroups table, column groupname. Leave NULL to use parameters instead.}

\item{stds}{A vector of standard names as listed in the EQWin eqstds table. Leave NULL to exclude standards. As these can apply to all stations standard values will be listed in a column to the left of the results.}

\item{stnStds}{TRUE/FALSE to include/exclude the station-specific standards listed in the eqstns table, column StnStd. As these are station-specific, the standard values will be listed in a comment linked to the station name in the table header.}

\item{date_approx}{An optional maximum number of days in the past or future to fetch results for stations where data is unavailable for the exact date.}

\item{save_path}{The path to save the Excel file. Default is "choose" to allow user to select a folder interactively.}

\item{dbPath}{The path to the EQWin database. Default is "X:/EQWin/WR/DB/Water Resources.mdb".}
}
\value{
An Excel workbook saved where requested.
}
\description{
Connection to EQWin is made using function \code{\link[=AccessConnect]{AccessConnect()}}
}
