% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSC_drainages.R
\name{WSC_drainages}
\alias{WSC_drainages}
\title{Concatenate WSC drainage polygons to single shapefile}
\usage{
WSC_drainages(
  inputs_folder = "choose",
  save_path = "choose",
  active_only = FALSE,
  limit_stns = list(stns_08 = c("08AB001", "08AC002", "08AC001", "08AA003", "08AA011",
    "08AA010", "08AA009", "08AA005", "08AA012", "08AA007", "08AA008"), stns_10 =
    c("10MB004", "10MA003", "10MA005", "10MA002", "10MA004", "10MA001", "10MB003",
    "10MC002", "10MC007", "10MD003", "10MD001", "10ED001", "10DB001", "10BE001",
    "10BE009", "10AC005", "10EB013", "10AA001", "10AA005", "10AA006", "10AD002",
    "10AA004", "10AB001"))
)
}
\arguments{
\item{inputs_folder}{The location of the folder containing folders for each drainage region of interest (01-11, as defined by the WSC). Defaults to "choose", which opens file explorer to let you select the location. NOTE: This should contain one or more folders named ONLY with the two-digit drainage region code, i.e. folder names 09, 10, etc.}

\item{save_path}{The path to the folder in which you wish to save the resultant ESRI shapefile. For hydrology this is likely G:/water/Hydrology/Data/BasinPolygons; parameter defaults to 'choose'. WARNING: option 'choose' only works on Windows, and some late-build R versions have a bug that prevents it from working every time.}

\item{active_only}{TRUE to retain only active stations; default FALSE retains discontinued stations as well.}

\item{limit_stns}{Specify stations if you wish to narrow the geographic scope of the output. Useful if your end-use is only within a section of a drainage region and you wish to limit file size. Format: one list element with a name of form stns_XX composed of a vector of station IDs for each drainage region you wish to limit. Preset for Yukon, set to NULL to use all stations present in the inputs_folder}
}
\value{
Two ESRI shapefiles (points + polygons) saved in the directory you specified.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Takes drainage polygons generated by the Water Survey of Canada and combines them into one large shapefile.

WSC drainage polygons are published with one point and one polygon shapefile per folder, one folder for each WSC station, and each WSC station folder is in turn in a folder for the larger drainage basin numbered 01 to 11. This is not convenient for most purposes, and this function facilitates further geospatial work by lumping all polygons into one shapefile.. It is designed to work with WSC products such as those found \href{https://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/HydrometricNetworkBasinPolygons/}{here}.
}
\seealso{
SAiVE::drainageBasins() to generate drainage polygons for any user-specified pour point.
}
