% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xle_processing.R
\name{xle_processing}
\alias{xle_processing}
\title{Solinst xle logger file processing}
\usage{
xle_processing(
  file,
  aq_upload = TRUE,
  master_sheet =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/1_YOWN_MASTER_TABLE/YOWN_MASTER.xlsx",
  logger_tracking =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/3_OTHER/YOWN_Logger_Tracking.xlsx",
  dropbox =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/9_LOGGER_FILE_DROPBOX",
  repo =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/1_YOWN_SITES/1_ACTIVE_WELLS"
)
}
\arguments{
\item{file}{Full path to the .xle file. This must be an XLE file from a Solinst logger.}

\item{aq_upload}{Logical, whether to upload data to Aquarius. If FALSE a data.frame of the data is returned. Uses function \code{\link[=aq_upload]{aq_upload()}} with default parameters, so you'll want to make sure your Aquarius login details are in your .Renviron file.}

\item{master_sheet}{Path to YOWN master sheet (.xlsx), used to match logger data to YOWN ID and ensure integrity.}

\item{logger_tracking}{Path to YOWN logger tracking sheet (.xlsx), used to track logger metadata. This part is done within a tryCatch block to enable this function to run even if the logger tracking sheet is open, missing, or otherwise inaccessible.}

\item{dropbox}{Path to YOWN logger dropbox folder, where the xle file is located. This is used to move the file to the appropriate folder after processing.}

\item{repo}{Path to YOWN Active Wells folder, which will be concatenated with the YOWN ID to create the final resting place for the xle file. If NULL the file will not get moved.}
}
\value{
Moves YOWN xle file to backups folder and appropriate YOWN Active Wells folder. Uploads data to Aquarius after performing unit checks and conversions if aq_upload is TRUE, else returns a data.frame.
}
\description{
Parse xle (Solinst) logger files, upload to Aquarius, sort into respective folder, and track logger metadata. Created to work within the Yukon Water Resources Branch's folder structure.
}
