% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedbackApp.R
\name{YGwater_feedback}
\alias{YGwater_feedback}
\title{Launch the YGwater feedback application}
\usage{
YGwater_feedback(
  db_path = "feedback.sqlite",
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port"),
  server = FALSE
)
}
\arguments{
\item{db_path}{Path to the SQLite database file where feedback will be stored.
Defaults to "feedback.sqlite" in the current working directory.}

\item{host}{Host address. Defaults to the value of \code{getOption('shiny.host', '127.0.0.1')}.}

\item{port}{Port number. Defaults to \code{getOption('shiny.port')}.}

\item{server}{Set to TRUE to run on Shiny Server, otherwise FALSE to run locally.}
}
\value{
Opens the feedback Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This application collects feedback from internal users about the YGwater app.
Feedback is stored in a local SQLite database.
}
