% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWN_solinst_telem.R
\name{YOWN_solinst_telem}
\alias{YOWN_solinst_telem}
\title{YOWN Solinst LevelSender Telemetry Processing}
\usage{
YOWN_solinst_telem(
  active_telem = c("YOWN-2201S"),
  path = "G:/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/9_XLE_file_dropbox/",
  auth_json =
    "G:\\\\water\\\\Groundwater\\\\2_YUKON_OBSERVATION_WELL_NETWORK\\\\4_YOWN_DATA_ANALYSIS\\\\1_WATER LEVEL\\\\00_AUTOMATED_REPORTING\\\\04_TELEMETRY\\\\gmail_interface.json"
)
}
\arguments{
\item{active_telem}{Character vector of YOWN sites for processing of SOLINST telemetry data.}

\item{path}{The path to the folder in which to save the .xle file.}

\item{auth_json}{The path to the .json file used for connection to gmail.}
}
\value{
Creates an .xle file in the folder specified by parameter \code{path}.
}
\description{
This script searches for new messages from the YOWN telemetry gmail inbox by specified sites with active telemetry, then downloads messages and combines each day's report into an XLE file which is then places in the XLE-dropbox folder for processing as usual. Trashed email upon successful creation of a .xle file.
}
