% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWN_WDC_Update.R
\name{YOWN_WDC_Update}
\alias{YOWN_WDC_Update}
\title{YOWN Water Data Catalogue Update Function}
\usage{
YOWN_WDC_Update(WDC_struc = yown_wdc_path(), master_file = yown_master_path())
}
\arguments{
\item{WDC_struc}{Link to master directory containing folders linked to the WDC.}

\item{master_file}{Path to YOWN master sheet.}
}
\value{
Updated files in the WDC folder structure
}
\description{
YOWN Water Data Catalogue Update Function
}
