% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_RawData.R
\name{YOWNplot_RawData}
\alias{YOWNplot_RawData}
\title{YOWN raw data writing}
\usage{
YOWNplot_RawData(
  AQID,
  timeSeriesID = "Wlevel_bgs.Calculated",
  saveTo = "desktop",
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{AQID}{Identity of YOWN site in the following format: "YOWN-XXXX" or "YOWN-XXXXD"}

\item{timeSeriesID}{Identity of the time series exactly as written in Aquarius (eg."Wlevel_bgs.Calculated")}

\item{saveTo}{Location for data files to be saved. Will create directory if it doesn't exist. Defaults to user's desktop.}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details. Passed to \code{\link[=aq_download]{aq_download()}}.}

\item{server}{The URL for your organization's Aquarius web server. Default is for the Yukon Water Resources Branch. Passed to \code{\link[=aq_download]{aq_download()}}.}
}
\value{
Writes .pdf of the full period of record for a specified YOWN site and time series ID, as well as a .csv containing all the raw data and copies a grade key to interpret data grading
}
\description{
This function downloads YOWN level data from Aquarius, screens data graded below "C", writes the csv file containing the entire period of record, and copies the grading key from the master location to the SaveTo directory
}
\details{
To store login credentials in your .renviron profile, run usethis::edit_r_environ() and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
