% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_SiteCompare.R
\name{YOWNplot_SiteCompare}
\alias{YOWNplot_SiteCompare}
\title{YOWN site comparative plot generation, in m bgs}
\usage{
YOWNplot_SiteCompare(
  YOWNindex,
  tsunit,
  chartRange = "all",
  chartXInterval = "1 month",
  saveTo = "desktop",
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{YOWNindex}{Character vector of YOWN site IDs (eg. c("YOWN-2201S", "YOWN-2201D", "YOWN-2202"))}

\item{tsunit}{Desired timeseries for plotting, CHOOSE FROM: "btoc", "bgs", "asl"}

\item{chartRange}{???}

\item{chartXInterval}{????}

\item{saveTo}{Location for data files to be saved as a character vector. Default "desktop" is to a new folder on your desktop; "choose" lets you interactively choose your save location.}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details. Passed to \code{\link[=aq_download]{aq_download()}}.}

\item{server}{Defaults to Yukon Water Resources Branch Aquarius web server. Passed to \code{\link[=aq_download]{aq_download()}}.}
}
\value{
Writes .pdf plot of WSC and YOWN data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plots multiple YOWN stations on the same chart
}
\details{
To store login credentials in your .renviron profile, call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
