% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWNplot_WSC.R
\name{YOWNplot_WSC}
\alias{YOWNplot_WSC}
\title{YOWN-WSC comparative plot generation relative to a common datum}
\usage{
YOWNplot_WSC(
  YOWNindex,
  WSCindex,
  title = "YOWN wells vs. WSC Sites",
  chartXInterval = "1 month",
  saveTo = "desktop",
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = "https://yukon.aquaticinformatics.net/AQUARIUS"
)
}
\arguments{
\item{YOWNindex}{Character vector of YOWN site IDs (eg. c("YOWN-2201S", "YOWN-2201D", "YOWN-2202"))}

\item{WSCindex}{Character vector of WSC site IDs (eg. c("09AB004", "09AH001"))}

\item{title}{The title you wish to see on the plot as character vector.}

\item{chartXInterval}{????}

\item{saveTo}{Location for data files to be saved as a character vector. Default "desktop" is to a new folder on your desktop; "choose" lets you interactively choose your save location.}

\item{login}{Your Aquarius login credentials as a character vector of two (eg. c("cmfische", "password") Default pulls information from your .renviron profile; see details. Passed to \code{\link[=aq_download]{aq_download()}}.}

\item{server}{The URL for your organization's Aquarius web server. Default is for the Yukon Water Resources Branch. Passed to \code{\link[=aq_download]{aq_download()}}.}
}
\value{
Writes .pdf plot of WSC and YOWN data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plots YOWN data alongside WSC stations, in m asl. Only works for stations that have values in the "Elevation" field in their Aquarius information page.
}
\details{
To store login credentials in your .renviron profile, call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS = "your password".
}
