% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Run the YGwater API}
\usage{
api(host = "0.0.0.0", port = 8000)
}
\arguments{
\item{host}{The host address to bind the server to. Default is '0.0.0.0'.}

\item{port}{The port number to listen on. Default is 8000.}
}
\value{
Runs the API server.
}
\description{
Launches the YGwater API using the plumber package.
}
\examples{
\dontrun{
run_api()
}
}
