% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Run the YGwater API}
\usage{
api(
  version = NULL,
  host = "127.0.0.1",
  port = 8000,
  server = "/water-data/api",
  dbName = "aquacache",
  dbHost = "localhost",
  dbPort = 5432
)
}
\arguments{
\item{version}{The API version to use. Default is NULL, which uses the latest version.}

\item{host}{The host address to bind the server to. Default is the local host.}

\item{port}{The port number to listen on. Default is 8000.}

\item{server}{The server path for the API. Default is '/water-data/api'.}

\item{dbName}{The name of the PostgreSQL database. Default is 'aquacache'.}

\item{dbHost}{The host address of the PostgreSQL database. Default is 'localhost'.}

\item{dbPort}{The port number of the PostgreSQL database. Default is 5432.}
}
\value{
Runs the API server.
}
\description{
Launches the YGwater API using the plumber package.
}
\examples{
\dontrun{
api()
}
}
