% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_download_curves.R
\name{aq_download_curves}
\alias{aq_download_curves}
\title{Get flow rating curves from Aquarius}
\usage{
aq_download_curves(
  loc_id,
  inputParameter = NULL,
  outputParameter = NULL,
  start = "1950-01-01",
  end = Sys.Date(),
  login = Sys.getenv(c("AQUSER", "AQPASS")),
  server = Sys.getenv("AQSERVER")
)
}
\arguments{
\item{loc_id}{The location ID, exactly as visible in Aquarius web portal, as a character vector of length 1. Typically of form \verb{29EA001} or \code{YOWN-0804}. Note that the timeseries id is not used, only the location ID.}

\item{inputParameter}{The input parameter for which you want to know the rating curve. If NULL, all rating curves for the location will be returned. This should match the 'Parameter' in Aquarius exactly.}

\item{outputParameter}{The output parameter for which you want to know the rating curve. If NULL, all rating curves for the location will be returned. This should match the 'Parameter' in Aquarius exactly.}

\item{start}{The first day or instant for which you want information, in UTC 0 timezone. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the first moment of the day. Times requested prior to the actual timeseries start will be adjusted to match available data.}

\item{end}{The last day or instant for which you want information, in UTC 0. You can specify a Date object, POSIXct object, or character vector of form yyyy-mm-dd or yyyy-mm-dd HH:mm:ss. If specifying date or POSIXct objects, the timezone attribute will be ignored. If only a date is specified it will be assigned the last moment of the day. Times requested prior to the actual timeseries end will be adjusted to match available data.}

\item{login}{Your Aquarius login credentials as a character vector of two. Default pulls information from your .renviron file; see details.}

\item{server}{The URL for your organization's Aquarius web server. Default pulls from your .renviron file; see details.}
}
\value{
A list with an element per rating model (one per inputParameter/outputParameter combination). Each element contains a list of rating curves, each with a data.frame of the relationship between input and output, the best-fit equation, and a plot of the data and fit. Metadata about the location is also included.
}
\description{
Fetches and processes rating curve data downloaded from an Aquarius web-hosted server and returns it in a concise format. Information about all rating curves for a location can be returned, or the user can specify a particular input and output parameter for which to retrieve the rating curve. This function also requests input-output relationships from Aquarius and fits a power law or polynomial model to the data; the best-fit model is selected based on AIC and the equation returned.
}
\details{
To store login credentials in your .renviron file, call \code{\link[usethis:edit]{usethis::edit_r_environ()}} and enter your username and password as value pairs, as AQUSER="your username" and AQPASS="your password". The server should be entered at server="your_server_url".
}
