% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{AquaConnect}
\alias{AquaConnect}
\title{Connect to the AquaCache database}
\usage{
AquaConnect(
  name = "AquaCache",
  host = Sys.getenv("AquaCacheHost"),
  port = Sys.getenv("AquaCachePort"),
  username = Sys.getenv("AquaCacheUser"),
  password = Sys.getenv("AquaCachePass"),
  RLS_user = Sys.getenv("RLS_user"),
  RLS_pass = Sys.getenv("RLS_pass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form AquaCacheHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter=value pair of form AquaCachePort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter=value pair of form AquaCacheUser="username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter=value pair of form AquaCachePass="password".}

\item{RLS_user}{Username for row-level security. Default searches the .Renviron file for parameter=value pair of form RLS_user="username" and if not found either prompts the user for input or logs in as 'public' if not interactive.}

\item{RLS_pass}{Password for row-level security. Default searches the .Renviron file for parameter=value pair of form RLS_pass="password" and if not found either prompts the user for input or logs in as 'public' if not interactive.}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the production version AquaCache database. For the development version see \code{\link[=hydrometConnect]{hydrometConnect()}}.
Database superusers will have access to all database records, but other users will be asked to provide their username and password to access records other than the 'public' ones. Note that this is \emph{not necessarily} the same username and password as the one used to log into the database itself.
}
