% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaConnect.R
\name{AquaConnect}
\alias{AquaConnect}
\title{Connect to the aquacache}
\usage{
AquaConnect(
  name = "aquacache",
  host = Sys.getenv("aquacacheHost"),
  port = Sys.getenv("aquacachePort"),
  username = Sys.getenv("aquacacheUser"),
  password = Sys.getenv("aquacachePass"),
  silent = FALSE
)
}
\arguments{
\item{name}{Database name.}

\item{host}{Database host address. By default searches the .Renviron file for parameter:value pair of form aquacacheHost="hostname".}

\item{port}{Connection port. By default searches the .Renviron file for parameter=value pair of form aquacachePort="1234".}

\item{username}{Username. By default searches the .Renviron file for parameter=value pair of form aquacacheUser="username". Refrain from using username with write privileges unless you absolutely know what you're doing.}

\item{password}{Password. By default searches the .Renviron file for parameter=value pair of form aquacachePass="password".}

\item{silent}{TRUE suppresses messages except for errors.}
}
\value{
A connection to the database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function exists to facilitate connecting to the aquacache database. You can pass your own database connection parameters, for example to connect to a dev instance instead of the default production instance.
Database superusers or admins will have access to all database records (unless your instance of aquacache was set up differently), but other users will be asked to provide their row level security username and password to access records other than the 'public' ones. Note that this is \emph{not necessarily} the same username and password as the one used to log into the database itself.
}
\details{
An attribute is added to the connection object to track if a transaction is active. This can be used by functions to determine if a transaction is already open, in which case functions can forgo opening a new transaction and instead use the existing one.
}
