% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinPrecip.R
\name{basinPrecip}
\alias{basinPrecip}
\title{Accumulated precipitation}
\usage{
basinPrecip(
  location,
  start = Sys.time() - 60 * 60 * 24,
  end = Sys.time(),
  map = FALSE,
  maptype = "dynamic",
  title = TRUE,
  raster_col = terra::map.pal("elevation", 20),
  silent = FALSE,
  con = NULL,
  hrdpa_loc = NULL,
  hrdps_loc = NULL,
  .raster_cache = NULL,
  .return_result_and_cache = FALSE
)
}
\arguments{
\item{location}{The location above which you wish to calculate precipitation. Specify either a WSC or WSC-like station ID (e.g. \code{"09AB004"}) for which there is a database entry, or coordinates in signed decimal degrees in form latitude, longitude (\code{"60.1234 -139.1234"}; note the space, negative sign, and lack of comma). Multiple locations are also supported as a character vector/list of WSC IDs, or as a list/matrix/data.frame of coordinate pairs (latitude, longitude).}

\item{start}{The start of the time period over which to accumulate precipitation. Use format \code{"yyyy-mm-dd hh:mm"} (character vector) in UTC time, or a POSIXct object (e.g. \code{Sys.time()-60*60*24} for one day in the past). In the case of a POSIXct object, the timezone is converted to UTC without modifying the time. See details if requesting earlier than 30 days prior to now.}

\item{end}{The end of the time period over which to accumulate precipitation. Other details as per \code{start}}

\item{map}{Should a map be output to the console? See details for more info.}

\item{maptype}{Choose from 'dynamic' or 'static'. Leaflet maps are interactive and return the entire Yukon territory with the requested basin highlighted. Static maps are printed to the console and show only the requested basin. Only used if map = TRUE.}

\item{title}{Should a title be added to the map? Only used if map = TRUE.}

\item{raster_col}{The color palette to use for the raster plot. Default is terra::map.pal("elevation", 10), \strong{not} the terra::plot default. In all cases a very light blue-grey is the lowest precip amount color. Refer to \code{\link[terra:map.pal]{terra::map.pal()}} for other color options. Only used if map = TRUE.}

\item{silent}{If TRUE, no output is printed to the console.}

\item{con}{A connection to the aquacache database. NULL uses \code{\link[=AquaConnect]{AquaConnect()}} and automatically disconnects.}

\item{hrdpa_loc}{The directory (folder) where past precipitation rasters are to be downloaded. Suggested use is to specify a repository where all such rasters are saved to speed processing time and reduce data usage. If using the default NULL, rasters will not persist beyond your current R session.}

\item{hrdps_loc}{The directory (folder) where forecast precipitation rasters are to be downloaded. A folder will be created for the specific parameter (in this case, precipitation) or selected if already existing.}
}
\value{
For a single location, returns a list containing precipitation statistics and optional map output. For multiple locations, returns a list of per-location outputs in the same order as the input. Output includes \code{raster_time_vector_UTC}, the vector of raster timestamps used.
}
\description{
Calculates accumulated precipitation above defined basins or immediately surrounding a specified point. If possible, make sure to specify a location in parameters hrdpa_loc and hrdps_loc to speed things up! Uses functions \code{\link[=getHRDPA]{getHRDPA()}} and \code{\link[=getHRDPS]{getHRDPS()}} to bring in precipi rasters if any need to be downloaded.

If specifying \code{location} as coordinates rather than a WSC station (or non WSC station, see note below), the numeric result will be of precipitation in the 2.5 km2 cell on which the point falls. Map output, if requested, will be for the smallest available drainage polygon containing that point.

Raster images of precipitation are only kept by ECCC for 30 days. If specifying a \code{start} or \code{end} prior to that without having precipitation rasters available locally and specified with \code{hrdpa_loc}, the start date/time will be adjusted and a warning given.

Rasters must be downloaded for each 6-hour period between \code{start} and \code{end}. It is therefore strongly suggested that you designate and use a folder for this purpose. \code{getHRDPA} is called for downloading and clipping rasters.

Additional spatial data is added to the maps when possible, Keeping in mind that large vector files can be lengthy for R to graphically represent. To ease this limitation, the watercourses file is left out when the drainage extent is greater than 30 000 km2; by that size, major water courses are expected to be represented by polygons in the waterbodies layer anyways.
}
