% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeReport.R
\name{bridgeReport}
\alias{bridgeReport}
\title{Bridge radar reporting utility - internal}
\usage{
bridgeReport(
  con = hydrometConnect(),
  locations = c("29AH001", "09AH005", "29AE007", "29AB011", "29AB010"),
  zoom = TRUE,
  zoom_days = 30,
  save_path = "choose"
)
}
\arguments{
\item{con}{A connection to the database. Default uses function \code{\link[=hydrometConnect]{hydrometConnect()}} with default settings.}

\item{locations}{The list of locations for which you want a distance measurement. These must either be reporting radar distance in Aquarius as Distance.Corrected or in the WRB database as type distance.}

\item{zoom}{Set TRUE if you want a zoomed-in plot.}

\item{zoom_days}{Set the number of days on the x-axis of the zoomed in plot.}

\item{save_path}{The path to the directory (folder) where the report should be saved. Default "choose" lets you select your folder, otherwise enter the path as a character string.}
}
\value{
A Microsoft Word report containing bridge radar information, saved in the location specified.
}
\description{
This function generates a report of distance between the water surface and bridges or other important infrastructure. The output is a Microsoft Word document on a Yukon Government template.
}
