% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_period.R
\name{calculate_period}
\alias{calculate_period}
\title{Calculate periodicity of data and add a column}
\usage{
calculate_period(data, timeseries_id, con = hydrometConnect(silent = TRUE))
}
\arguments{
\item{data}{The data.frame or data.table for which to calculate periodicity. Must contain at minimum a column named 'datetime' (in POSIXct format) with no missing values.}

\item{timeseries_id}{The ID of the timeseries for which to calculate periodicity. Used to fetch any data points lacking a period, as well as to search for additional data points if there are too few to calculate a period in the provided \code{data}.}

\item{con}{A connection to the database, created with \code{\link[DBI:dbConnect]{DBI::dbConnect()}} or using the utility function \code{\link[=hydrometConnect]{hydrometConnect()}}.}
}
\value{
A data.frame or data.table with calculated periods as ISO8601 formatted strings in a column named 'period'.
}
\description{
Utility function to calculate periodicity of continuous-type temporal data and prepate a column named 'period' with ISO8601 formatted periods. Will identify changes to periodicity within data, for example moving from 1-hour intervals to 6-hour intervals.
}
