% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSnowTemplate.R
\name{createSnowTemplate}
\alias{createSnowTemplate}
\title{Creates snow survey template for specific survey target date and circuit}
\usage{
createSnowTemplate(
  target_date,
  circuits = "all",
  save_path = "choose",
  snowCon = NULL
)
}
\arguments{
\item{target_date}{The target date of the snow survey, given in the form yyyy-mm-dd as a text string or Date object. Example: for the march snow survey of 2024, target_date = '2024-03-01'}

\item{circuits}{The circuit or circuits for which we are creating a snow survey template. Options are Alaska, Carmacks, Dawson, Haines Junction, Kluane National Park, Mayo, North Slope, Old Crow, Pelly Farm, Ross River, Southern Lakes, Teslin, Watson Lake, Whitehorse, YEC. "all" will create a workbook for all circuits listed in the Snow database. Default is "all".}

\item{save_path}{The path where you want the circuit workbook(s) saved. Default "choose" lets you pick your folder.}

\item{snowCon}{A snow database connection object. If left NULL the function will attempt to connect to the snow database using function \code{\link[=snowConnect]{snowConnect()}} and close the connection when finished.}
}
\value{
A snow survey template for the specified circuit(s) and target date. This excel workbook has a sheet for every snow course within the circuit and contains a summary sheet with current and previous years stats.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

NOTE: locations must be associated with a circuit in the snow database for this function to work. If you are missing locations, you will need to add them to the snow database before running this function.

This function creates snow survey templates for a snow survey circuit in an excel workbook. The function takes the master template and copies it into multiple sheets with circuit name, course name and snow survey target date pre-filled. The outputted template is meant to be filled out by those conducting snow surveys.

You need to be running this function from a computer with access to the Snow database (and probably production instance) for full functionality. Note that this is a change as of March 2025, as the function used to work without a connection to the database with limited functionality. This change was done in order to use the database's set snow course groupings.
}
