% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_processing.R
\name{html_processing}
\alias{html_processing}
\title{html logger file processing}
\usage{
html_processing(
  html_file,
  master =
    "G:/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/1_YOWN_MASTER_TABLE/YOWN_MASTER.xlsx",
  logger_tracking =
    "G:/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/3_OTHER/YOWN_Logger_Tracking.xlsx",
  YOWN_logger_folder =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/9_LOGGER_FILE_DROPBOX"
)
}
\arguments{
\item{html_file}{File name of .html file, must be in the logger dropbox folder}
}
\value{
Uploaded file to Aquarius and logger files sorted
}
\description{
Parse YOWN html logger files, upload to Aquarius, sort into respective folder, and track logger metadata. Used "Pressure" column after converting to m water column. If pressure column does not exist, converts from depth using formula supplied by In Situ
}
