% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_processing.R
\name{html_processing}
\alias{html_processing}
\title{In Situ html logger file processing}
\usage{
html_processing(
  file,
  aq_upload = TRUE,
  master_sheet = yown_master_path(),
  logger_tracking = yown_tracking_path(),
  dropbox = yown_dropbox_path(),
  repo = yown_active_path()
)
}
\arguments{
\item{file}{Full path to the .html file. This must be an html file from an In Situ logger.}

\item{aq_upload}{Logical, whether to upload data to Aquarius. If FALSE a data.frame of the data is returned. Uses function \code{\link[=aq_upload]{aq_upload()}} with default parameters, so you'll want to make sure your Aquarius login details are in your .Renviron file.}

\item{master_sheet}{Path to YOWN master sheet (.xlsx), used to match logger data to YOWN ID and ensure integrity. Defaults to a WRB network location.}

\item{logger_tracking}{Path to YOWN logger tracking sheet (.xlsx), used to track logger metadata. Defaults to a WRB network location. This part is done within a tryCatch block to enable this function to run even if the logger tracking sheet is open, missing, or otherwise inaccessible.}

\item{dropbox}{Path to YOWN logger dropbox folder, where the html file is located. Defaults to a WRB network location. This is used to move the file to the appropriate folder after processing.}

\item{repo}{Path to YOWN Active Wells folder, which will be concatenated with the YOWN ID to create the final resting place for the html file. Defaults to a WRB network location. If NULL the file will not get moved.}
}
\value{
Moves YOWN html file to backups folder and appropriate YOWN Active Wells folder. Uploads data to Aquarius if \code{aq_upload = TRUE} after performing unit checks and conversions. Adds pressure column if missing, based on depth column and conversions calcs on In Situ website.
}
\description{
Parse YOWN html logger files, upload to Aquarius, sort into respective folder, and track logger metadata. Used "Pressure" column after converting to m water column. If pressure column does not exist, converts from deth using formula supplied by In Situ
}
