% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_processing.R
\name{html_processing}
\alias{html_processing}
\title{In Situ html logger file processing}
\usage{
html_processing(
  file,
  aq_upload = TRUE,
  master_sheet =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/1_YOWN_MASTER_TABLE/YOWN_MASTER.xlsx",
  logger_tracking =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/2_SPREADSHEETS/3_OTHER/YOWN_Logger_Tracking.xlsx",
  dropbox =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/9_LOGGER_FILE_DROPBOX",
  repo =
    "//env-fs/env-data/corp/water/Groundwater/2_YUKON_OBSERVATION_WELL_NETWORK/1_YOWN_SITES/1_ACTIVE_WELLS"
)
}
\arguments{
\item{file}{Full path to the .html file. This must be an html file from an In Situ logger.}

\item{aq_upload}{Logical, whether to upload data to Aquarius. If FALSE a data.frame of the data is returned.}

\item{master_sheet}{Path to YOWN master sheet (.xlsx), used to match logger data to YOWN ID and ensure integrity.}

\item{logger_tracking}{Path to YOWN logger tracking sheet (.xlsx), used to track logger metadata. This part is done within a tryCatch block to enable this function to run even if the logger tracking sheet is open, missing, or otherwise inaccessible.}

\item{dropbox}{Path to YOWN logger dropbox folder, where the html file is located. This is used to move the file to the appropriate folder after processing.}

\item{repo}{Path to YOWN Active Wells folder, which will be concatenated with the YOWN ID to create the final resting place for the html file. If NULL the file will not get moved.}
}
\value{
Moves YOWN html file to backups folder and appropriate YOWN Active Wells folder. Uploads data to Aquarius after performing unit checks and conversions. Adds pressure column if missing, based on depth column and conversions calcs on In Situ website
#' @export
}
\description{
Parse YOWN html logger files, upload to Aquarius, sort into respective folder, and track logger metadata. Used "Pressure" column after converting to m water column. If pressure column does not exist, converts from deth using formula supplied by In Situ
}
