% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnR.R
\name{learnR_answer}
\alias{learnR_answer}
\title{Answers for the learnR lessons}
\usage{
learnR_answer(lesson, part = 1)
}
\arguments{
\item{lesson}{The lesson number for which you want the answer.}

\item{part}{The part of the lesson for which you want the answer.}
}
\value{
The answer for the specified lesson, both as text and as a command to be executed in the console.
}
\description{
Answers for the learnR lessons
}
\examples{

learnR_answer(1)
}
