% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnR.R
\name{learnR_hint}
\alias{learnR_hint}
\title{Hints for the learnR lessons}
\usage{
learnR_hint(lesson, part = 1)
}
\arguments{
\item{lesson}{The lesson number for which you want a hint.}

\item{part}{The part of the lesson for which you want a hint.}
}
\value{
A hint for the specified lesson.
}
\description{
Hints for the learnR lessons
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

learnR_hint(1)
\dontshow{\}) # examplesIf}
}
