% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiscrete.R
\name{plotDiscrete}
\alias{plotDiscrete}
\title{Discrete (lab or field point data) plotting function}
\usage{
plotDiscrete(
  start,
  end = Sys.Date() + 1,
  locations = NULL,
  locGrp = NULL,
  parameters = NULL,
  paramGrp = NULL,
  log = FALSE,
  facet_on = "params",
  loc_code = FALSE,
  rows = "auto",
  target_datetime = TRUE,
  colorblind = FALSE,
  lang = "en",
  dbSource = "EQ",
  dbPath = "//carver/infosys/EQWin/WR/DB/Water Resources.mdb"
)
}
\arguments{
\item{start}{The date to fetch data from, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'.}

\item{end}{The end date to fetch data up to, passed as a Date, POSIXct, or character vector of form 'yyyy-mm-dd HH:MM'. Dates and character vectors are converted to POSIXct with timezone 'MST'. Default is the current date.}

\item{locations}{A vector of station names or codes. If dbSource == 'AC': from AquaCache 'locations' table use column 'location', 'name', or 'name_fr' (character vector) or 'location_id' (numeric vector). If dbSource == 'EQ' use EQWiN 'eqstns' table, column 'StnCode' or leave NULL to use \code{locGrp} instead.}

\item{locGrp}{Only used if \code{dbSource} is 'EQ'. A station group as listed in the EWQin 'eqgroups' table, column 'groupname.' Leave NULL to use \code{locations} instead.}

\item{parameters}{A vector of parameter names or codes. If dbSource == 'AC': from AquaCache 'parameters' table use column 'param_name' or 'param_name_fr' (character vector) or 'parameter_id' (numeric vector). If dbSource == 'EQ' use EQWin 'eqparams' table, column 'ParamCode' or leave NULL to use \code{paramGrp} instead.}

\item{paramGrp}{Only used if \code{dbSource} is 'EQ'. A parameter group as listed in the EQWin 'eqgroups' table, column 'groupname.' Leave NULL to use \code{parameters} instead.}

\item{log}{Should the y-axis be log-transformed?}

\item{facet_on}{Should the plot be faceted by locations or by parameters? Specify one of 'locs' or 'params'. Default is 'locs'.}

\item{loc_code}{Should the location code be used instead of the full location name?}

\item{rows}{The number of rows to use in the facet grid. Default is 'auto' to automatically determine the number of rows based on the number of facets.}

\item{target_datetime}{Should the plot datetime use the 'target' datetime instead of the 'actual' datetime? Default is TRUE. This is only applicable is dbSource == 'AC'.}

\item{colorblind}{Should the plot be colorblind-friendly? Default is FALSE.}

\item{lang}{The language to use for the plot. Currently only "en" and "fr" are supported. Default is "en", and this is only supported for dbSource == 'AC'.}

\item{dbSource}{The database source to use, 'AC' for AquaCache or 'EQ' for EQWin. Default is 'EQ'. Connections to AquaCache are made using function \code{\link[=AquaConnect]{AquaConnect()}}.}

\item{dbPath}{The path to the EQWin database, if called for in parameter \code{dbSource}. Default is "//carver/infosys/EQWin/WR/DB/Water Resources.mdb".}
}
\value{
A zoomable plot of the data from EQWin.
}
\description{
Plots data directly from EQWin or from the AquaCache for one or more location (station) and one or more parameter. Depending on the setting for argument 'facet_on', the function can either make a facet plot where each station is a facet (with parameters as distinct traces) or where each parameter is a facet (with locations as distinct traces). Values above or below the detection limit are shown as the detection limit but symbolized differently (open circles).
}
