% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiTimeseries.R
\name{plotMultiTimeseries}
\alias{plotMultiTimeseries}
\title{Plot multiple timeseries from the aquacache}
\usage{
plotMultiTimeseries(
  type = "traces",
  timeseries_ids = NULL,
  locations = NULL,
  sub_locations = NULL,
  parameters = NULL,
  record_rates = NULL,
  aggregation_types = NULL,
  z = NULL,
  z_approx = NULL,
  start_date = Sys.Date() - 1,
  end_date = Sys.Date(),
  lead_lag = NULL,
  log = FALSE,
  invert = NULL,
  slider = FALSE,
  shareX = TRUE,
  shareY = FALSE,
  datum = TRUE,
  title = TRUE,
  custom_title = NULL,
  filter = NULL,
  unusable = FALSE,
  historic_range = NULL,
  lang = "en",
  line_scale = 1,
  axis_scale = 1,
  legend_scale = 1,
  legend_position = "v",
  gridx = FALSE,
  gridy = FALSE,
  webgl = TRUE,
  resolution = NULL,
  tzone = "auto",
  data = FALSE,
  con = NULL,
  as_of = NULL,
  stats_period = "full"
)
}
\arguments{
\item{type}{Are you looking for multiple traces on one plot ('traces') or multiple subplots ('subplots')? Default is 'traces'.}

\item{timeseries_ids}{Vector of timeseries IDs to plot, if known (else leave NULL). When provided, location/parameter inputs are ignored.}

\item{locations}{The location or locations for which you want a plot. If specifying multiple locations matched to the parameters and record_rates 1:1. The location:parameter combos must be in the local database.}

\item{sub_locations}{Your desired sub-locations, if applicable. Default is NULL as most locations do not have sub-locations. Specify as the exact name of the sub-locations (character) or the sub-location IDs (numeric). Matched one to one to the locations and parameters or recycled if specified as length one.}

\item{parameters}{The parameter or parameters you wish to plot. You can specify parameter names (text) or id (numeric) from table 'parameters'. If specifying multiple parameters matched to the locations and record_rates 1:1. The location:parameter combos must be in the local database.}

\item{record_rates}{The recording rate for the parameters and locations. In most cases there are not multiple recording rates for a location and parameter combo and you can leave this NULL. Otherwise NULL will default to the most frequent record rate, or you can set this as one of '< 1 day', '1 day', '1 week', '4 weeks', '1 month', 'year'. Matched one to one to the locations and parameters or recycled if specified as length one. Can be passed in a character string or number of seconds coercible to a period by \code{\link[lubridate:period]{lubridate::period()}}.}

\item{aggregation_types}{The period type(s) for the parameter and location to plot. Options other than the default NULL are 'sum', 'min', 'max', or '(min+max)/2', which is how the daily 'mean' temperature is often calculated for meteorological purposes (you can also specify the numeric entry from the aggregation_types AquaCache table). NULL will search for what's available and get the first timeseries found in this order: 'instantaneous', followed by the 'mean', '(min+max)/2', 'min', and 'max' in that order. Matched one to one to the locations and parameters or recycled if specified as length one.}

\item{z}{Depth/height in meters further identifying the timeseries of interest. Default is NULL, and where multiple elevations exist for the same location/parameter/record_rate/aggregation_type combo the function will default to the absolute elevation value closest to ground. Otherwise set to a numeric value. Matched one to one to the locations and parameters or recycled if specified as length one.}

\item{z_approx}{Number of meters by which to approximate the elevation. Default is NULL, which will use the exact elevation. Otherwise set to a numeric value. Matched one to one to the locations and parameters or recycled if specified as length one.}

\item{start_date}{The day or datetime on which to start the plot as character, Date, or POSIXct. Date-only values start at the beginning of that day in \code{tzone}. Default is one year ago.}

\item{end_date}{The day or datetime on which to end the plot as character, Date, or POSIXct. Date-only values include the full selected day in \code{tzone}. Default is today.}

\item{lead_lag}{The number of \strong{hours} to lead or lag the data. Default is NULL, which will not lead or lag any of the timeseries, otherwise set to a signed numeric value. Matched one to one to the locations and parameters. Not used for \code{type} 'subplots'.}

\item{log}{Should any/all y axes use a logarithmic scale? Specify as a logical (TRUE/FALSE) vector of length 1 or of length equal to the number of traces you wish to plot. Default is FALSE.}

\item{invert}{Should the y-axis be inverted? TRUE/FALSE, or leave as NULL to use the database default. Specify as logical vector of same length as 'locations' and 'parameters', or a single value that gets recycled for all. Default is NULL.}

\item{slider}{Should a slider be included to show where you are zoomed in to? If TRUE the slider will be included but this prevents horizontal zooming or zooming in using the box tool. If legend_position is set to 'h', slider will be set to FALSE due to interference. Default is FALSE.}

\item{shareX}{Should the x-axis be shared across facets? Default is TRUE (dates are shared).}

\item{shareY}{Should the y-axis be shared across facets? Default is FALSE (values are not shared).}

\item{datum}{Should a vertical offset be applied to the data? Looks for it in the database and applies it if it exists, only for water level and distance. Default is TRUE.}

\item{title}{Should a title be included? TRUE/FALSE.}

\item{custom_title}{Custom title to be given to the plot. Default is NULL, which will set the title as the location name as entered in the database.}

\item{filter}{Should an attempt be made to filter out spurious data? Will calculate the rolling IQR and filter out clearly spurious values. Set this parameter to an integer, which specifies the rolling IQR 'window'. The greater the window, the more effective the filter but at the risk of filtering out real data. Negative values are always filtered from parameters "water level" ("niveau d'eau"), "flow" ("d\\u00E9bit"), "snow depth" ("profondeur de la neige"), "snow water equivalent" ("\\u00E9quivalent en eau de la neige"), "distance", and any "precip" related parameter. Otherwise all values below -100 are removed.}

\item{unusable}{Should unusable data be displayed? Default is FALSE. Note that unusable data is not used in the calculation of historic ranges.}

\item{historic_range}{Default NULL will plot historic ranges for up to two traces and no historical ranges at all if more than two traces; TRUE will plot historic ranges for all traces; FALSE will not plot historic ranges.}

\item{lang}{The language to use for the plot. Currently only "en" and "fr" are supported. Default is "en".}

\item{line_scale}{A scale factor to apply to the size (width) of the lines. Default is 1.}

\item{axis_scale}{A scale factor to apply to the size of axis labels. Default is 1.}

\item{legend_scale}{A scale factor to apply to the size of text in the legend. Default is 1.}

\item{legend_position}{The position of the legend, 'v' for vertical on the right side or 'h' for horizontal on the bottom. Default is 'v'. If 'h', slider will be set to FALSE due to interference.}

\item{gridx}{Should gridlines be drawn on the x-axis? Default is FALSE}

\item{gridy}{Should gridlines be drawn on the y-axis? Default is FALSE}

\item{webgl}{Use WebGL ("scattergl") for faster rendering when possible. Set
to FALSE to force standard scatter traces.}

\item{resolution}{The resolution at which to plot the data. Default is NULL, which will adjust for reasonable plot performance depending on the date range. Otherwise set to one of "max", "hour", "day".}

\item{tzone}{The timezone to use for the plot. Default is "auto", which will use the system default timezone. Otherwise set to a valid timezone string.}

\item{data}{Should the data used to create the plot be returned? Default is FALSE.}

\item{con}{A connection to the target database. NULL uses \code{\link[=AquaConnect]{AquaConnect()}} and automatically disconnects.}

\item{as_of}{Optional point-in-time timestamp at which measurement values and
stored daily summaries should be reconstructed. Character, Date, and
POSIXct inputs are supported. Date-like inputs are interpreted as the end
of that day in \code{tzone}. When \code{NULL} (default), current data are used.}

\item{stats_period}{Historical range statistics period. Use "full" for
full-record statistics or "30yr" for the most recent 30-year statistics
when the connected database provides them. Default is "full".}
}
\value{
A plotly object and a data frame with the data used to create the plot (if \code{data} is TRUE).
}
\description{
This function plots multiple continuous timeseries from the aquacache using either a facet plot or a single plot with multiple traces. The plot is zoomable and hovering over the historical ranges or the measured values brings up additional information. If corrections are applied to the data within AquaCache, the corrected values will be used.
}
