% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverlap.R
\name{plotOverlap}
\alias{plotOverlap}
\title{Plot of overlapping years or portions thereof}
\usage{
plotOverlap(
  location,
  sub_location = NULL,
  parameter,
  record_rate = NULL,
  startDay = 1,
  endDay = 365,
  tzone = "MST",
  years = NULL,
  datum = TRUE,
  title = TRUE,
  custom_title = NULL,
  invert = NULL,
  slider = TRUE,
  filter = NULL,
  unusable = FALSE,
  historic_range = "last",
  rate = "day",
  line_scale = 1,
  axis_scale = 1,
  legend_scale = 1,
  legend_position = "v",
  hover = TRUE,
  gridx = FALSE,
  gridy = TRUE,
  lang = "en",
  data = FALSE,
  con = NULL
)
}
\arguments{
\item{location}{The location for which you want a plot.}

\item{sub_location}{Your desired sub-location, if applicable. Default is NULL as most locations do not have sub-locations. Specify as the exact name of the sub-location (character) or the sub-location ID (numeric).}

\item{parameter}{The parameter name (text) or code (numeric) you wish to plot. The location:parameter combo must be in the local database.}

\item{record_rate}{The recording rate for the parameter and location to plot. In most cases there are not multiple recording rates for a location and parameter combo and you can leave this NULL. Otherwise NULL will default to the most frequent record rate.}

\item{startDay}{The start day of year for the plot x-axis. Can be specified as a number from 1 to 365, as a character string of form "yyyy-mm-dd", or as a date object. Either way the day of year is the only portion used, specify years to plot under parameter \code{years}.}

\item{endDay}{The end day of year for the plot x-axis. As per \code{startDay}.}

\item{tzone}{The timezone to use for graphing. Only really evident for a small number of days.}

\item{years}{The years to plot. If \code{startDay} and \code{endDay} cover December 31 - January 1, select the December year(s). Max 10 years, NULL = current year.}

\item{datum}{Should a vertical datum be applied to the data, if available? TRUE or FALSE.}

\item{title}{Should a title be included?}

\item{custom_title}{Custom title to be given to the plot. Default is NULL, which will set the title as the location name as entered in the database.}

\item{invert}{Should the y-axis be inverted? TRUE/FALSE, or leave as default NULL to use the database default value.}

\item{slider}{Should a slider be included to show where you are zoomed in to? If TRUE the slider will be included but this prevents horizontal zooming or zooming in using the box tool. If legend_position is set to 'h', slider will be set to FALSE due to interference. Default is TRUE.}

\item{filter}{Should an attempt be made to filter out spurious data? Will calculate the rolling IQR and filter out clearly spurious values. Set this parameter to an integer, which specifies the rolling IQR 'window'. The greater the window, the more effective the filter but at the risk of filtering out real data. Negative values are always filtered from parameters "water level" ("niveau d'eau"), "flow" ("débit"), "snow depth" ("profondeur de la neige"), "snow water equivalent" ("équivalent en eau de la neige"), "distance", and any "precip" related parameter. Otherwise all values below -100 are removed.}

\item{unusable}{Should unusable data be displayed? Default is FALSE. Note that unusable data is not used in the calculation of historic ranges.}

\item{historic_range}{Should the historic range parameters be calculated using all available data (i.e. from start to end of records) or only up to the last year specified in "years"? Choose one of "all" or "last".}

\item{rate}{The rate at which to plot the data. Default is NULL, which will adjust for reasonable plot performance depending on the date range. Otherwise set to one of "max", "hour", "day".}

\item{line_scale}{A scale factor to apply to the size (width) of the lines. Default is 1.}

\item{axis_scale}{A scale factor to apply to the size of axis labels. Default is 1.}

\item{legend_scale}{A scale factor to apply to the size of text in the legend. Default is 1.}

\item{legend_position}{The position of the legend, 'v' for vertical on the right side or 'h' for horizontal on the bottom. Default is 'v'. If 'h', slider will be set to FALSE due to interference.}

\item{hover}{Should hover text be included? Default is TRUE.}

\item{gridx}{Should grid lines be drawn on the x-axis? Default is FALSE}

\item{gridy}{Should grid lines be drawn on the y-axis? Default is FALSE}

\item{lang}{The language to use for the plot. Currently only "en" and "fr" are supported. Default is "en".}

\item{data}{Should the data used to create the plot be returned? Default is FALSE.}

\item{con}{A connection to the target database. NULL uses AquaConnect from this package and automatically disconnects.}
}
\value{
An html plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A pared-down, sped up, simplified version of the original \code{\link[=ggplotOverlap]{ggplotOverlap()}} function. Created for use with a simple Shiny application linked to the Flood Atlas, but will likely see other uses.
}
