% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeseries.R
\name{plotTimeseries}
\alias{plotTimeseries}
\title{Plot a continous timeseries from the aquacache}
\usage{
plotTimeseries(
  timeseries_id = NULL,
  location = NULL,
  sub_location = NULL,
  parameter = NULL,
  record_rate = NULL,
  aggregation_type = NULL,
  z = NULL,
  z_approx = NULL,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  invert = NULL,
  slider = TRUE,
  datum = FALSE,
  title = TRUE,
  custom_title = NULL,
  filter = NULL,
  unusable = FALSE,
  grades = FALSE,
  approvals = FALSE,
  qualifiers = FALSE,
  historic_range = TRUE,
  lang = "en",
  line_scale = 1,
  axis_scale = 1,
  legend_scale = 1,
  legend_position = "v",
  hover = TRUE,
  gridx = FALSE,
  gridy = FALSE,
  webgl = TRUE,
  resolution = NULL,
  tzone = "auto",
  raw = FALSE,
  imputed = FALSE,
  data = FALSE,
  build_plot = TRUE,
  con = NULL,
  as_of = NULL,
  stats_period = "full"
)
}
\arguments{
\item{timeseries_id}{The timeseries ID to plot, if known (else leave NULL).}

\item{location}{The location for which you want a plot.}

\item{sub_location}{Your desired sub-location, if applicable. Default is NULL as most locations do not have sub-locations. Specify as the exact name of the sub-location (character) or the sub-location ID (numeric).}

\item{parameter}{The parameter name (text) or code (numeric) that you wish to plot. The location:sublocation:parameter combo must be in the local database.}

\item{record_rate}{The recording rate for the parameter and location to plot, from column 'record_rate' of table 'timeseries'. In most cases there are not multiple recording rates for a location and parameter combo and you can leave this NULL. Otherwise NULL will default to the most frequent record rate. Can be passed in a character string or number of seconds coercible to a period by \code{\link[lubridate:period]{lubridate::period()}}.}

\item{aggregation_type}{The period type for the parameter and location to plot. Options other than the default NULL are 'sum', 'min', 'max', or '(min+max)/2', which is how the daily 'mean' temperature is often calculated for meteorological purposes. NULL will search for what's available and get the first timeseries found in this order: 'instantaneous', followed by the 'mean', '(min+max)/2', 'min', and 'max'.}

\item{z}{Depth/height in meters further identifying the timeseries of interest. Default is NULL, and where multiple elevations exist for the same location/parameter/record_rate/aggregation_type combo the function will default to the absolute elevation value closest to ground. Otherwise set to a numeric value.}

\item{z_approx}{Number of meters by which to approximate the elevation. Default is NULL, which will use the exact elevation. Otherwise set to a numeric value.}

\item{start_date}{The day or datetime on which to start the plot as character, Date, or POSIXct. Date-only values start at the beginning of that day in \code{tzone}. Default is one year ago.}

\item{end_date}{The day or datetime on which to end the plot as character, Date, or POSIXct. Date-only values include the full selected day in \code{tzone}. Default is today.}

\item{invert}{Should the y-axis be inverted? TRUE/FALSE, or leave as default NULL to use the database default value.}

\item{slider}{Should a slider be included to show where you are zoomed in to? If TRUE the slider will be included but this prevents horizontal zooming or zooming in using the box tool. If legend_position is set to 'h', slider will be set to FALSE due to interference. Default is TRUE.}

\item{datum}{Should a vertical offset be applied to the data? Looks for it in the database and applies it if it exists. Default is FALSE.}

\item{title}{Should a title be included?}

\item{custom_title}{Custom title to be given to the plot. Default is NULL, which will set the title as the location name as entered in the database.}

\item{filter}{Should an attempt be made to filter out spurious data? Will calculate the rolling IQR and filter out clearly spurious values. Set this parameter to an integer, which specifies the rolling IQR 'window'. The greater the window, the more effective the filter but at the risk of filtering out real data. Negative values are always filtered from parameters "water level" ("niveau d'eau"), "flow" ("débit"), "snow depth" ("profondeur de la neige"), "snow water equivalent" ("équivalent en eau de la neige"), "distance", and any "precip" related parameter. Otherwise all values below -100 are removed.}

\item{unusable}{Should unusable data be displayed? Default is FALSE. Note that unusable data is not used in the calculation of historic ranges.}

\item{grades}{Should grades be included on the y-axis? Default is FALSE.}

\item{approvals}{Should approvals be included on the y-axis? Default is FALSE.}

\item{qualifiers}{Should qualifiers be included on the y-axis? Default is FALSE.}

\item{historic_range}{Should the historic range be plotted? Default is TRUE.}

\item{lang}{The language to use for the plot. Currently only "en" and "fr" are supported. Default is "en".}

\item{line_scale}{A scale factor to apply to the size (width) of the line. Default is 1.}

\item{axis_scale}{A scale factor to apply to the size of axis labels. Default is 1.}

\item{legend_scale}{A scale factor to apply to the size of text in the legend. Default is 1.}

\item{legend_position}{The position of the legend, 'v' for vertical on the right side or 'h' for horizontal on the bottom. Default is 'v'. If 'h', slider will be set to FALSE due to interference.}

\item{hover}{Should hover text be included? Default is TRUE.}

\item{gridx}{Should gridlines be drawn on the x-axis? Default is FALSE}

\item{gridy}{Should gridlines be drawn on the y-axis? Default is FALSE}

\item{webgl}{Use WebGL ("scattergl") for faster rendering when possible. Set to FALSE to force standard scatter traces.}

\item{resolution}{The resolution at which to plot the data. Default is NULL, which will adjust for reasonable plot performance depending on the date range. Otherwise set to one of "max", "hour", "day".}

\item{tzone}{The timezone to use for the plot. Default is "auto", which will use the system default timezone. Otherwise set to a valid timezone string or a numeric UTC offset (in hours).}

\item{raw}{Should raw data be used instead of corrected data (if corrections exist)? Default is FALSE.}

\item{imputed}{Should imputed data be displayed differently? Default is FALSE.}

\item{data}{Should the data used to create the plot be returned? Default is FALSE.}

\item{build_plot}{Should a plotly object be constructed? Internal optimisation for Shiny modules. Default is TRUE.}

\item{con}{A connection to the target database. NULL uses \code{\link[=AquaConnect]{AquaConnect()}} and automatically disconnects.}

\item{as_of}{Optional point-in-time timestamp at which measurement values,
daily summaries, quality-control metadata, and unusable-grade exclusions
should be reconstructed. Character, Date, and POSIXct inputs are supported.
Date-like inputs are interpreted as the end of that day in \code{tzone}. When
\code{NULL} (default), current data are used.}

\item{stats_period}{Historical range statistics period. Use "full" for
full-record statistics or "30yr" for the most recent 30-year statistics
when the connected database provides them. Default is "full".}
}
\value{
A plotly object and a data frame with the data used to create the plot (if \code{data} is TRUE).
}
\description{
This function plots continuous timeseries from the aquacache. The plot is zoomable and hovering over the historical ranges or the measured values brings up additional information. If corrections are applied to the data within AquaCache, the corrected values will be used.
}
