% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowBulletinStats.R
\name{snowBulletinStats}
\alias{snowBulletinStats}
\title{Function for creating snow bulletin statistics - internal}
\usage{
snowBulletinStats(
  year,
  month,
  basins = NULL,
  excel_output = FALSE,
  save_path = "choose",
  synchronize = FALSE,
  source = "aquacache",
  con = NULL
)
}
\arguments{
\item{year}{Year for which the snow bulletin stats are calculated.}

\item{month}{Month for which the snow bulletin stats are calculated. Options are 3, 4 or 5 as a vector of length 1.}

\item{basins}{The name of the sub_basin you wish to generate stats for. One or more of "Upper Yukon", "Teslin", "Central Yukon", "Pelly", "Stewart", "White", "Lower Yukon", "Porcupine", "Peel", "Liard", "Alsek" as a character vector, or NULL for all basins. North Slope will be added when AquaCache is updated with the new snow database.}

\item{excel_output}{If TRUE, calculated stats will be output in multiple excel tables.}

\item{save_path}{The path to the directory (folder) where the excel files should be saved. Enter the path as a character string or leave as 'choose' to select interactively. Not used if excel_output = FALSE.}

\item{synchronize}{Should the timeseries be synchronized with source data? If TRUE, all timeseries used in this function will be synchronized. If FALSE (default), none will be synchronized. Currently is a placeholder and does not work.}

\item{source}{Should the SWE statistics for stations (station_stats) be calculated from the 'aquacache' database or the 'snow' database? Default is 'aquacache'.}

\item{con}{A connection to the database. Leave as NULL to use function \code{\link[=AquaConnect]{AquaConnect()}} with default settings and close the connection automatically afterwards.}
}
\value{
A list of data.frames OR, if requested, three Excel workbooks.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function generates the statistics associated with the snow bulletin for a specified date. The output is a list of tables (in R), or a number of excel workbooks. Data is fetched from the local AquaCache database created/maintained by the AquaCache package OR from the snow database.
}
\details{
To download data, you MUST have your aquacache credentials added to your .Renviron file. See function \code{\link[=AquaConnect]{AquaConnect()}} or talk to the database administrator/data scientist for help.
}
