% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowInfo.R
\name{snowInfo}
\alias{snowInfo}
\title{Plots and tabular data for snow survey locations}
\usage{
snowInfo(
  locations = "all",
  inactive = FALSE,
  save_path = "choose",
  stats = TRUE,
  complete_yrs = TRUE,
  plots = TRUE,
  plot_type = "combined",
  quiet = FALSE,
  con = NULL
)
}
\arguments{
\item{locations}{The list of locations requested, as a character vector of length n. Default "all" fetches all stations.}

\item{inactive}{Boolean specifying whether to include inactive stations. In this case, a station with no measurements for 5 or more years is considered inactive. Default is FALSE, i.e. no inactive stations.}

\item{save_path}{The path where the .csv(s) and plots should be saved. Set to NULL for data only as an R object. Plots are not created if there is no save path.}

\item{stats}{set TRUE if you want basic statistics (mean, min, max) and calculated trends.}

\item{complete_yrs}{Should only years with complete data be used? If TRUE, will not use the current year's data until after May.}

\item{plots}{Set TRUE if you want plots of SWE, depth, and density generated (but see next parameter).}

\item{plot_type}{Set to "separate" for 3 plots per location, or "combined" for a single compound plot per location.}

\item{quiet}{Suppresses most messages and warnings.}

\item{con}{A connection to the snow database. Leave as NULL to use \code{\link[=AquaConnect]{AquaConnect()}} to establish a connection, which will be closed when finished. If you pass your own connection remember to close it when done.}
}
\value{
A list with four data.frames: location metadata, basic statistics, trend information, and snow course measurements is returned to the R environment. In addition, an Excel workbook is saved to the save_path with the four data.frames, and a new folder created to hold SWE and depth plots for each station requested.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is intended to facilitate the reporting of snow survey data by compiling basic statistics (years of record, missing years, mean, max, etc.), trend information (Mann-Kendall direction and p-value, Sen's slope), and creating simple plots of SWE, depth, and density for all requested stations. At its most basic (parameters to FALSE or NULL where applicable), the result is a list of two data.frames to the R environment with location metadata and field measurements.
}
\seealso{
\code{\link[=waterInfo]{waterInfo()}} for a similar function dealing with water flow/level.
}
