% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamLine.R
\name{streamLine}
\alias{streamLine}
\title{StreamLine Shiny application for WRB data viewing}
\usage{
streamLine()
}
\value{
Opens a Shiny application.
}
\description{
Loads a Shiny application for viewing WRB data.
}
\examples{
streamLine()
}
