% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularReport.R
\name{tabularReport}
\alias{tabularReport}
\title{Tabular output of hydrometric data}
\usage{
tabularReport(
  level_locations = "all",
  flow_locations = "all",
  snow_locations = "all",
  bridge_locations = "all",
  precip_locations = "default",
  report_datetime = Sys.time(),
  past = 28,
  save_path = NULL,
  archive_path = NULL,
  log_level = "INFO",
  con = NULL
)
}
\arguments{
\item{level_locations}{List of water level locations to include in the report, as a character vector. "default" is a pre-determined list of locations across the territory, "all" fetches all level reporting locations in the DB. NULL will not create the table.}

\item{flow_locations}{List of flow locations to include in the report, as a character vector. "default" is a pre-determined list of locations across the territory. "all" fetches all flow reporting locations in the DB. NULL will not create the table.}

\item{snow_locations}{List of snow pillow locations to include in the report, as a character vector. "default" includes all of the WRB snow pillows as of Feb 2023, "all" fetches all snow pillow locations in the DB. NULL will not create the table.}

\item{bridge_locations}{List of bridge freeboard radar locations to include in the report, as a character vector. "default" includes all of the radars as of Feb 2023, "all" fetches all snow pillow locations in the DB. NULL will not create the table.}

\item{precip_locations}{List of flow/level locations for which to report precipitation. "default" is a pre-determined list of locations, "all" is all locations for which there is a drainage polygon (which may be more or less than the number of stations reporting level or flow information). NULL will not create the table. WARNING: this portion of the script is slow. Setting this parameter to "all" could take about an hour to get all information together.}

\item{report_datetime}{Date-time for which to generate the report. Defaults to the current time.}

\item{past}{The number of days in the past for which you want data. Will be rounded to yield table columns covering at least one week, at most 4 weeks. 24, 28, and 72 hour change columns are always rendered.}

\item{save_path}{The path where you wish to save the Excel workbook. A folder will be created for each day's report. 'choose' will bring up a file dialog to select the folder if the session is interactive. Default is 'choose'.}

\item{archive_path}{The path to yesterday's file, if you wish to include yesterday's comments in this report. Full path, including extension .xlsx. Function expects a workbook exactly as produced by this function, plus of course the observer comments. Default is 'choose', set to NULL to not use a previous report.}

\item{log_level}{Logging threshold for file logging. One of "DEBUG", "INFO", "WARN", or "ERROR". Default is "INFO".}

\item{con}{A connection to the aquacache database. NULL uses \code{\link[=AquaConnect]{AquaConnect()}} and automatically disconnects.}
}
\value{
The path to which the report was saved, and Excel workbook containing the report with one tab per timeseries type.
}
\description{
Creates a report of hydrometric, snow pack, precipitation, and air temperature conditions in Excel format, each table on a separate tab. List of stations/locations can be user-defined if desired. Connection is established using AquaConnect by default and MUST connect to a database created and maintained by the package AquaCache.

Note that data can only be as recent as the last incorporation to the database. If you need the most up to date data possible, run AquaCache::getNewContinuous first.
}
