% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularReport.R
\name{tabularReport}
\alias{tabularReport}
\title{Tabular output of hydrometric data}
\usage{
tabularReport(
  con = hydrometConnect(silent = TRUE),
  level_locations = "all",
  flow_locations = "all",
  snow_locations = "all",
  bridge_locations = "all",
  precip_locations = "default",
  past = 28,
  save_path = "choose",
  archive_path = "choose"
)
}
\arguments{
\item{con}{A connection to the database. Default uses function \code{\link[=hydrometConnect]{hydrometConnect()}} with default settings.}

\item{level_locations}{List of water level locations to include in the report, as a character vector. "default" is a pre-determined list of locations across the territory, "all" fetches all level reporting locations in the DB. NULL will not create the table.}

\item{flow_locations}{List of flow locations to include in the report, as a character vector. "default" is a pre-determined list of locations across the territory. "all" fetches all flow reporting locations in the DB. NULL will not create the table.}

\item{snow_locations}{List of snow pillow locations to include in the report, as a character vector. "default" includes all of the WRB snow pillows as of Feb 2023, "all" fetches all snow pillow locations in the DB. NULL will not create the table.}

\item{bridge_locations}{List of bridge freeboard radar locations to include in the report, as a character vector. "default" includes all of the radars as of Feb 2023, "all" fetches all snow pillow locations in the DB. NULL will not create the table.}

\item{precip_locations}{List of flow/level locations for which to report precipitation. "default" is a pre-determined list of locations, "all" is all locations for which there is a drainage polygon (which may be more or less than the number of stations reporting level or flow information). NULL will not create the table. WARNING: this portion of the script is slow. Setting this parameter to "all" could take about an hour to get all information together.}

\item{past}{The number of days in the past for which you want data. Will be rounded to yield table columns covering at least one week, at most 4 weeks. 24, 28, and 72 hour change columns are always rendered.}

\item{save_path}{The path where you wish to save the Excel workbook. A folder will be created for each day's report.}

\item{archive_path}{The path to yesterday's file, if you wish to include yesterday's comments in this report. Full path, including exension .xlsx. Function expects a workbook exactly as produced by this function, plus of course the observer comments.}
}
\value{
An Excel workbook containing the report with one tab per timeseries type.
}
\description{
Creates a report of hydrometric, snow pack, and precipitation conditions in Excel format, each table on a separate tab. List of stations/locations can be user-defined if desired. Connection is established using hydrometConnect by default and MUST connect to a database created and maintained by the package HydroMetDB.
Note that data can only be as recent as the last incorporation to the database. If you need the most up to date data possible, run HydroMetDB::getNewContinuous first.
}
