% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YGwater-package.R
\docType{package}
\name{YGwater-package}
\alias{YGwater}
\alias{YGwater-package}
\title{YGwater: Utility functions for the Yukon Water Science and Stewardship Branch}
\description{
Stores various functions for use across the Water Science and Stweardship Branch disciplines, including for generating reports, gathering and processing data, and creating plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YukonWRB/YGwater}
  \item Report bugs at \url{https://github.com/YukonWRB/YGWater/issues}
}

}
\author{
\strong{Maintainer}: Ghislain de Laplante \email{ghislaindel@gmail.com} (\href{https://orcid.org/0000-0002-5093-9185}{ORCID})

Authors:
\itemize{
  \item Cole Fischer \email{cole.fischer@hotmail.com} (\href{https://orcid.org/0000-0002-7266-5760}{ORCID})
  \item Emilie Stewart-Jones \email{emilie.stewart-jones@yukon.ca}
}

}
\keyword{internal}
