% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YGwater.R
\name{YGwater}
\alias{YGwater}
\title{Shiny application for internal WRB use}
\usage{
YGwater(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port"),
  dbName = "aquacache",
  dbHost = Sys.getenv("aquacacheHost"),
  dbPort = Sys.getenv("aquacachePort"),
  dbUser = Sys.getenv("aquacacheUser"),
  dbPass = Sys.getenv("aquacachePass"),
  accessPath =
    "//env-fs/env-data/corp/water/Data/Databases_virtual_machines/databases/EQWinDB",
  server = FALSE
)
}
\arguments{
\item{host}{Host address. Leave default to run locally, set to "0.0.0.0" to enable others to connect. Depends on the port specified in \code{port} to be open on the host machine. Ignored if \code{server} is set to TRUE (Shiny Server takes care of that).}

\item{port}{Port number (numeric) on which to serve the app. Leave default to use the default port specified in your user options. The port you specify must be open on the host machine for it to broadcast to the network. Ignored if \code{server} is set to TRUE (Shiny Server takes care of that).}

\item{dbName}{Name of the aquacache database. Default is "aquacache".}

\item{dbHost}{Host address of the aquacache database. Default is pulled from the .Renviron file.}

\item{dbPort}{Port number of the aquacache database. Default is pulled from the .Renviron file.}

\item{dbUser}{Username for the aquacache database. Default is pulled from the .Renviron file.}

\item{dbPass}{Password for the aquacache database. Default is pulled from the .Renviron file.}

\item{accessPath}{Path to the folder where EQWin databases are stored. Default is "//env-fs/env-data/corp/water/Data/Databases_virtual_machines/databases/EQWinDB". The function will search for all *.mdb files in this folder (but not its sub-folders) and list them as options.}

\item{server}{Set to TRUE to run on Shiny Server, otherwise FALSE to run locally.}
}
\value{
Opens a Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This application is designed for use by WRB or other YG staff, and is designed to allow greater flexibility and tune ability of plots and outputs than the public facing app (Streamline). In addition, certain portions of the application interface with databases other than aquacache such as EQWin or the snow database, and may allow for edit privileges to these databases. This app is also used to roll out new, experimental features before they are added to public applications
}
