% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YGwater.R
\name{YGwater}
\alias{YGwater}
\title{Shiny application for internal WRB use}
\usage{
YGwater(
  host = getOption("shiny.host", "127.0.0.1"),
  port = getOption("shiny.port")
)
}
\arguments{
\item{host}{Host address. Leave default to run locally, set to "0.0.0.0" to enable others to connect. Depends on the port specified in \code{port} to be open on the host machine.}

\item{port}{Port number (numeric). Leave default to use the default port specified in your user options. The port you specify must be open on the host machine for it to broadcast to the network.}
}
\value{
Opens a Shiny application.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This application is designed for use by WRB or other YG staff, and is designed to allow greater flexibility and tune ability of plots and outputs than the public facing app (Streamline). In addition, certain portions of the application interface with databases other than AquaCache such as EQWin or the snow database, and may allow for edit privileges to these databases. This app is also used to roll out new, experimental features before they are added to Streamline.
}
