% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YOWN_WDC_Update.R
\name{YOWN_WDC_Update}
\alias{YOWN_WDC_Update}
\title{YOWN Water Data Catalogue Update Function}
\usage{
YOWN_WDC_Update(
  WDC_struc =
    "\\\\\\\\envgeoserver\\\\share\\\\WaterResources\\\\Groundwater\\\\YOWN_DATA\\\\",
  master_sheet =
    "G:\\\\water\\\\Groundwater\\\\2_YUKON_OBSERVATION_WELL_NETWORK\\\\2_SPREADSHEETS\\\\1_YOWN_MASTER_TABLE\\\\YOWN_MASTER.xlsx"
)
}
\arguments{
\item{WDC_struc}{Link to master directory containing folders linked to the WDC}

\item{master_sheet}{Path to YOWN master sheet}
}
\value{
Updated files in the WDC folder structure
}
\description{
YOWN Water Data Catalogue Update Function
}
